### 2024-02-13 #6350 @kthyng
Hi @mdtanker and thanks for your submission! We have a backlog of submissions so I will add this to our waitlist. In the meantime, please suggest 5 reviewers from the database listed above or your own (non-conflicted) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them. Thanks for your patience.

However first can you comment on your readme note "This package is in early stages of development"? Your package looks relatively established, but this note seems to indicate it might be better to wait to publish in JOSS?
