### 2024-02-13 mdtanker/polartoolkit#168 @mdtanker
Update early-stages warning
Update the warning in the documentation about this being an early-stage package. Still include some warning about checking the changelog before updating the package to a new major version.

- [ ] update `readme.md`
- [ ] update `docs/index.md`
