### 2024-02-13 #6350 @mdtanker
Hi @kthyng, thanks for the instructions. I will find 5 reviewers and post them here.

I've had that warning there since I started the project, but it is probably time to remove it, or at least edit it. Since I know there are users of `PolarToolkit` who are new to programming, I am  just trying to warn them about the hazard of blindly updating to new major version without first checking the change log for breaking changes.

I've added an issue to fix this: https://github.com/mdtanker/polartoolkit/issues/168
