### 2024-02-13 #6350 @mdtanker
Suggested reviewers:
* Ken Mankoff: [mankoff](https://github.com/mankoff)
     * extensive research background based in Antarctica
* Wei Ji Leong: [weiji14](https://github.com/weiji14)
     * core developer at `PyGMT` (a core dependency of `PolarToolkit`)  and Antarctic researcher
     * he is mentioned in the acknowledgements, if that affects anything
* Jessica Scheick: [JessicaS11](https://github.com/JessicaS11)
     * cryosphere remote sensing and open-source software developer
* Chad Greene: [chadagreene](https://github.com/chadagreene)
     * not Python oriented, but created `Antarctic-Mapping-Tools`, which is a very similar package to `PolarToolkit`, but written in MatLab
* Penny How: [PennyHow](https://github.com/PennyHow)
     * open-source software developer focused on Greenland research
