### 2024-02-13 #6350 @mdtanker
> Wordcount for `paper.md` is 1532

I think it is closer to 800 words. There are lots of commented-out sections in the .md file that I think are being mistakingly included in that 1532.
