+++ docs/index.md
@@ -1,150 +1,3 @@
-# PolarToolkit
-
-Helpful tools for polar researchers
-
-```{include} ../README.md
-:start-after: <!-- SPHINX-START1 -->
-:end-before: <!-- SPHINX-END1 -->
+```{danger}
+Antarctic-Plots has been renamed to PolarToolkit, please use the updated docs <a href='https://polartoolkit.readthedocs.io/'>here</a>
 ```
-
-```{image} cover_fig.png
-:alt: cover figure
-:width: 400px
-:align: center
-```
-
-```{admonition} Early-stages of development
-:class: seealso
-
-This package is in early stages of development! This means that we are still adding a lot of new features and sometimes we make changes to the ones we already have while we try to improve the software based on users' experience, test new ideas, make better design decisions, etc. Some of these changes could be **backwards incompatible**. Keep that in mind before you update polartoolkit to a newer version.
-```
-
-
-```{include} ../README.md
-:start-after: <!-- SPHINX-START2 -->
-:end-before: <!-- SPHINX-END2 -->
-```
-
-
-```{admonition} How to contribute
-:class: seealso
-
-I really welcome all forms of contribution! If you have any questions, comments or suggestions, please open a [discussion]() or [issue (feature request)]() on the [GitHub page](https://github.com/mdtanker/polartoolkit/)!
-
-Also, please feel free to share how you're using PolarToolkit, I'd love to know.
-
-Please, read our [Contributor Guide](https://github.com/mdtanker/polartoolkit/blob/main/.github/CONTRIBUTING.md) to learn
-how you can contribute to the project.
-```
-
-```{note}
-*Many parts of this documentation was adapted from the* [Fatiando project](https://www.fatiando.org/).
-```
-
-```{toctree}
-:maxdepth: 1
-:hidden:
-overview
-install
-gallery/index.md
-tutorial/index.md
-```
-
-<!-- ```{nbgallery}
----
-caption: 📚 Gallery
-maxdepth: 1
-glob:
----
-gallery/basic_map
-gallery/extend_pygmt
-gallery/setting_projection
-gallery/3D_stack
-gallery/subplots
-gallery/subplot_layout
-``` -->
-
-<!-- ```{nbgallery}
----
-caption: 🚶 Tutorials
-maxdepth: 1
-glob:
----
-tutorial/fetch_walkthrough
-tutorial/profile_walkthrough
-tutorial/regions_walkthrough
-tutorial/utils_walkthrough
-``` -->
-
-```{toctree}
-:maxdepth: 1
-:hidden:
-:caption: Tips
-tips.ipynb
-```
-
-```{toctree}
-:maxdepth: 2
-:hidden:
-:caption: 📖 Reference documentation
-api/polartoolkit
-citing.md
-changelog.md
-references.rst
-```
-
-```{toctree}
-:maxdepth: 1
-:hidden:
-:caption: ℹ️ Other resources
-contributing.md
-Source code on GitHub <https://github.com/mdtanker/polartoolkit>
-```
-
-
-::::{grid} 2
-:::{grid-item-card} {octicon}`rocket` Getting started?
-:text-align: center
-New to PolarToolkit? Start here!
-```{button-ref} overview
-    :click-parent:
-    :color: primary
-    :outline:
-    :expand:
-```
-:::
-
-:::{grid-item-card} {octicon}`comment-discussion` Need help?
-:text-align: center
-Start a discussion on GitHub!
-```{button-link} https://github.com/mdtanker/polartoolkit/discussions
-    :click-parent:
-    :color: primary
-    :outline:
-    :expand:
-    Discussions
-```
-:::
-
-:::{grid-item-card} {octicon}`file-badge` Reference documentation
[…]
