+++ CHANGELOG.md
@@ -9,6 +9,12 @@ with commit parsing of [angular commits](https://github.com/angular/angular.js/b
 * switch GHA from hynek to build and twine ([`91c6ca9`](https://github.com/mdtanker/polartoolkit/commit/91c6ca9d11afd8197bf3fa043c7c8d2a5e9617d3))
 * explicitly include packages ([`a76a0e9`](https://github.com/mdtanker/polartoolkit/commit/a76a0e99c2aaec43b0a35218136609a970450b64))
 ### 🧰 Chores / Maintenance
+* mark test mass change as issue ([`a3bde43`](https://github.com/mdtanker/polartoolkit/commit/a3bde4376c336d82b1e600fdf7306e7cf580267b))
+* add issue marker to imagery and basal melt fetches ([`95bce69`](https://github.com/mdtanker/polartoolkit/commit/95bce69504dec80d3581b6ed864c9aaf892f70a9))
+* make fig an arg of plot_grd ([`e762a5a`](https://github.com/mdtanker/polartoolkit/commit/e762a5a7518932d5f630d7087517912055f3a0e3))
+* add kwarg arg to add_scalebar ([`df374ba`](https://github.com/mdtanker/polartoolkit/commit/df374ba838dea59c454cca2314a0b93841d1370c))
+* specify args in functions ([`9febc0e`](https://github.com/mdtanker/polartoolkit/commit/9febc0e46b3d405c7c83a12653344a59c3fb629b))
+* automatically updating changelog [skip ci] ([`c4c3eae`](https://github.com/mdtanker/polartoolkit/commit/c4c3eaecd266451eb4d8ac7d9b3b8fbec6f37235))
 * automatically updating changelog [skip ci] ([`248525e`](https://github.com/mdtanker/polartoolkit/commit/248525efe0e121676783067aa094f9fa756716cd))
 * remove dependa-bot commits from changelog ([`44a322c`](https://github.com/mdtanker/polartoolkit/commit/44a322c4059a69e9af735232799d6583f4b86e84))
 * automatically updating changelog [skip ci] ([`208dbd4`](https://github.com/mdtanker/polartoolkit/commit/208dbd4da9319361b876f05f49f80588141f9cc8))
@@ -22,6 +28,8 @@ with commit parsing of [angular commits](https://github.com/angular/angular.js/b
 * editing workflows ([`a059f50`](https://github.com/mdtanker/polartoolkit/commit/a059f50061eb8e4399e2aebed5b1491f44dd0e89))
 * collect usage stats ([`b5d16ac`](https://github.com/mdtanker/polartoolkit/commit/b5d16ac3792e3ef7791138274d1c069a108cc110))
 ### 📖 Documentation
+* fix disclaimer ([`8075024`](https://github.com/mdtanker/polartoolkit/commit/80750247c570b8d98774b02dcab64e33fd57d56c))
+* re-run all docs ([`0776879`](https://github.com/mdtanker/polartoolkit/commit/07768799052e77a59e36783bc2350e352d3ad4a7))
 * add pygmt and pip to RTD.env ([`9a8ae51`](https://github.com/mdtanker/polartoolkit/commit/9a8ae51b6cc51d9b798252e67932a05ec1148742))
 * back to conda for RTD ([`d2a4c1f`](https://github.com/mdtanker/polartoolkit/commit/d2a4c1f7e8d610fefd358917d61459d6b38939c0))
 * switch RTD from mamba to python ([`aeab560`](https://github.com/mdtanker/polartoolkit/commit/aeab560cc1872299c9be10e27f5f8f5327349091))
[…]
