+++ pyproject.toml
@@ -97,10 +97,10 @@ all = [
 
 [project.urls]
 Homepage = "https://github.com/mdtanker/polartoolkit"
-Documentation = "https://antarctic-plots.readthedocs.io/"
+Documentation = "https://polartoolkit.readthedocs.io/"
 "Bug Tracker" = "https://github.com/mdtanker/polartoolkit/issues"
 Discussions = "https://github.com/mdtanker/polartoolkit/discussions"
-Changelog = "https://antarctic-plots.readthedocs.io/en/latest/changelog.html"
+Changelog = "https://polartoolkit.readthedocs.io/en/latest/changelog.html"
 
 [tool.setuptools.packages.find]
 where = ["src"]
