+++ CHANGELOG.md
@@ -9,21 +9,14 @@ with commit parsing of [angular commits](https://github.com/angular/angular.js/b
 * switch GHA from hynek to build and twine ([`91c6ca9`](https://github.com/mdtanker/polartoolkit/commit/91c6ca9d11afd8197bf3fa043c7c8d2a5e9617d3))
 * explicitly include packages ([`a76a0e9`](https://github.com/mdtanker/polartoolkit/commit/a76a0e99c2aaec43b0a35218136609a970450b64))
 ### 🧰 Chores / Maintenance
-* automatically updating changelog [skip ci] ([`efa34a2`](https://github.com/mdtanker/polartoolkit/commit/efa34a23909fa0243c65c55abbc51c5c44addc4a))
+* ignore auto update changelog commits in changelog ([`01a092e`](https://github.com/mdtanker/polartoolkit/commit/01a092ee4cc77a60270deb2477de6cec70b002cd))
 * mark test mass change as issue ([`a3bde43`](https://github.com/mdtanker/polartoolkit/commit/a3bde4376c336d82b1e600fdf7306e7cf580267b))
 * add issue marker to imagery and basal melt fetches ([`95bce69`](https://github.com/mdtanker/polartoolkit/commit/95bce69504dec80d3581b6ed864c9aaf892f70a9))
 * make fig an arg of plot_grd ([`e762a5a`](https://github.com/mdtanker/polartoolkit/commit/e762a5a7518932d5f630d7087517912055f3a0e3))
 * add kwarg arg to add_scalebar ([`df374ba`](https://github.com/mdtanker/polartoolkit/commit/df374ba838dea59c454cca2314a0b93841d1370c))
 * specify args in functions ([`9febc0e`](https://github.com/mdtanker/polartoolkit/commit/9febc0e46b3d405c7c83a12653344a59c3fb629b))
-* automatically updating changelog [skip ci] ([`c4c3eae`](https://github.com/mdtanker/polartoolkit/commit/c4c3eaecd266451eb4d8ac7d9b3b8fbec6f37235))
-* automatically updating changelog [skip ci] ([`248525e`](https://github.com/mdtanker/polartoolkit/commit/248525efe0e121676783067aa094f9fa756716cd))
 * remove dependa-bot commits from changelog ([`44a322c`](https://github.com/mdtanker/polartoolkit/commit/44a322c4059a69e9af735232799d6583f4b86e84))
-* automatically updating changelog [skip ci] ([`208dbd4`](https://github.com/mdtanker/polartoolkit/commit/208dbd4da9319361b876f05f49f80588141f9cc8))
-* automatically updating changelog [skip ci] ([`3edbaeb`](https://github.com/mdtanker/polartoolkit/commit/3edbaeb2f52e4f76f19a649d22e5cff0e1cd7e54))
 * fix _static path warning ([`257f6e7`](https://github.com/mdtanker/polartoolkit/commit/257f6e728e6fd24d695b3b8ee5acc3140b260aa7))
-* automatically updating changelog [skip ci] ([`5dd0a40`](https://github.com/mdtanker/polartoolkit/commit/5dd0a40ba6289c35091615f443e74a9a7ce20af0))
-* automatically updating changelog [skip ci] ([`8f92bcb`](https://github.com/mdtanker/polartoolkit/commit/8f92bcb81b74afeec2391af30b63936199df4789))
-* automatically updating changelog [skip ci] ([`d200e9a`](https://github.com/mdtanker/polartoolkit/commit/d200e9a06d59106e75a6a93d4eff590b82814b43))
[…]
