+++ src/polartoolkit/__init__.py
@@ -9,4 +9,4 @@
 
 from __future__ import annotations
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
