+++ pyproject.toml
@@ -27,7 +27,7 @@ classifiers = [
     "Typing :: Typed",
 ]
 readme = "README.md"
-version = "0.2.1"
+version = "0.3.0"
 license = {file = "LICENSE"}
 
 keywords = ["cryosphere", "antarctica", "arctic", "greenland", "maps", "plotting", "pygmt", "cross-section"]
