+++ docs/index.md
@@ -33,7 +33,7 @@ I really welcome all forms of contribution! If you have any questions, comments 
 
 Also, please feel free to share how you're using PolarToolkit, I'd love to know.
 
-Please, read our [Contributor Guide](https://github.com/mdtanker/polartoolkit/blob/main/.github/CONTRIBUTING.md) to learn
+Please, read our [Contributor Guide](https://polartoolkit.readthedocs.io/en/latest/contributing.html) to learn
 how you can contribute to the project.
 ```
 
