+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 Automatically updated by
 [python-semantic-release](https://python-semantic-release.readthedocs.io/en/latest/)
 with commit parsing of [angular commits](https://github.com/angular/angular.js/blob/master/DEVELOPERS.md#commits).
+## Unreleased
+
+### 📖 Documentation
+* fix links to contrib guide ([`5c6af8d`](https://github.com/mdtanker/polartoolkit/commit/5c6af8de8ab984fa645f868ba23eadd492f0d4b7))
 
 ## v0.3.0 (2024-02-18)
 ### 📦️ Build
