### 2024-02-20 3fffd15
chore: add ignores to pre-commit
