+++ .pre-commit-config.yaml
@@ -16,8 +16,9 @@ exclude: |
       .github/ISSUE_TEMPLATE/bug_report.md|
       .github/ISSUE_TEMPLATE/feature_request.md|
       .github/config.yml|
-      .github/CONTRIBUTING.md|
+      CONTRIBUTING.md|
       binder/environment.yml|
+      templates/CHANGELOG.md.j2|
   )$
 
 repos:
