### 2024-02-20 f72c26c
chore: add test_fetch make command
