+++ .github/config.yml
@@ -5,7 +5,7 @@
 
 # Comment to be posted to on first time issues
 newIssueWelcomeComment: >
-  👋 Thanks for opening your first issue here! Please make sure you filled out the template with as much detail as possible. You might also want to take a look at our [contributing guidelines](https://github.com/mdtanker/polartoolkit/blob/main/docs/contribute.md).
+  👋 Thanks for opening your first issue here! Please make sure you filled out the template with as much detail as possible. You might also want to take a look at our [contributing guidelines](https://polartoolkit.readthedocs.io/en/latest/contributing.html).
 
 # Configuration for new-pr-welcome - https://github.com/behaviorbot/new-pr-welcome
 
@@ -13,7 +13,7 @@ newIssueWelcomeComment: >
 newPRWelcomeComment: >
   💖 Thanks for opening this pull request! 💖
 
-  Please make sure you read our [contributing guidelines](https://github.com/mdtanker/polartoolkit/blob/main/docs/contribute.md).
+  Please make sure you read our [contributing guidelines](https://polartoolkit.readthedocs.io/en/latest/contributing.html).
 
   A few things to keep in mind:
 
