### 2024-02-20 b412249
docs: update install instructions
