### 2024-02-20 12a684d
chore: add hashes to all fetch calls
