+++ src/polartoolkit/fetch.py
@@ -276,12 +276,17 @@ def sample_shp(name: str) -> str:
     str
         file path as a string
     """
+    if name == "Disco_deep_transect":
+        known_hash = "70e86b3bf9775dd824014afb91da470263edf23159a9fe34107897d1bae9623e"
+    elif name == "Roosevelt_Island":
+        known_hash = "83434284808d067b8b18b649e41287a63f01eb2ce581b2c34ee44ae3a1a5ca2a"
     path = pooch.retrieve(
         url=f"https://github.com/mdtanker/polartoolkit/raw/main/data/{name}.zip",
         path=f"{pooch.os_cache('pooch')}/polartoolkit/shapefiles",
         processor=pooch.Unzip(),
-        known_hash=None,
+        known_hash=known_hash,
     )
+
     val: str = next(p for p in path if p.endswith(".shp"))
     return val
 
@@ -337,7 +342,7 @@ def mass_change(
         url=url,
         fname=zip_fname,
         path=f"{pooch.os_cache('pooch')}/polartoolkit/ice_mass",
-        known_hash=None,
+        known_hash="8d09ffcce4e84fba8cabbc85ee79fec4de36419f8b242ca1f95adacaa2f229e3",
         progressbar=True,
         processor=pooch.Unzip(
             extract_dir="Smith_2020",
@@ -430,7 +435,7 @@ def basal_melt(variable: str = "w_b") -> typing.Any:
         url=url,
         fname=fname,
         path=f"{pooch.os_cache('pooch')}/polartoolkit/ice_mass/Admusilli_2020",
-        known_hash=None,
+        known_hash="c14f7059876e6808e3869853a91a3a17a776c95862627c4a3d674c12e4477d2a",
         progressbar=True,
         processor=preprocessing,
     )
@@ -570,7 +575,7 @@ def ice_vel(
         fname="measures_ice_vel_phase_map.nc",
         path=f"{pooch.os_cache('pooch')}/polartoolkit/ice_velocity",
         downloader=EarthDataDownloader(),
-        known_hash=None,
+        known_hash="fa0957618b8bd98099f4a419d7dc0e3a2c562d89e9791b4d0ed55e6017f52416",
         progressbar=True,
         processor=preprocessor,
     )
@@ -620,7 +625,7 @@ def modis_moa(
             path=f"{pooch.os_cache('pooch')}/polartoolkit/imagery",
             downloader=EarthDataDownloader(),
             processor=pooch.Decompress(method="gzip", name="moa125_2009_hp1_v02.0.tif"),
-            known_hash=None,
+            known_hash="101fa22295f94f6eab487d208c051cf81c9af925355b124a04e3d96463af5b72",
             progressbar=True,
         )
     elif version == "750m":
@@ -630,7 +635,7 @@ def modis_moa(
             path=f"{pooch.os_cache('pooch')}/polartoolkit/imagery",
             downloader=EarthDataDownloader(),
             processor=pooch.Decompress(method="gzip", name="moa750_2009_hp1_v02.0.tif"),
-            known_hash=None,
+            known_hash="90d1718ea0971795ec102482c47f308ba08ba2b88383facb9fe210877e80282c",
             progressbar=True,
         )
     else:
@@ -664,7 +669,7 @@ def imagery() -> str:
         fname="lima.zip",
         path=f"{pooch.os_cache('pooch')}/polartoolkit/imagery",
         processor=pooch.Unzip(),
-        known_hash=None,
+        known_hash="7e7daa7af128f1ad18ac597d95d716ba26f745e75f8abb81c10049419a070c37",
         progressbar=True,
     )
     return typing.cast(str, next(p for p in path if p.endswith(".tif")))
[…]
