### 2024-02-20 4ee6945
chore: update testing_env and add make command
