+++ env/test_env.yml
@@ -1,52 +1,0 @@
-# Requirements for running the GitHub Action: test
-# only need run-time dependencies
-name: polartoolkit
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - python>=3.9
-  - pip
-  # Run-time
-  - pandas
-  - openpyxl
-  - pooch
-  - tqdm
-  - verde>=1.8.0
-  - xarray
-  - netcdf4
-  - h5netcdf
-  - harmonica>=0.6.0
-  - pyproj
-  - matplotlib
-  - pyogrio
-  - rioxarray
-  - scipy
-  - numpy
-  - pygmt
-  - geopandas
-  - zarr
-  - python-dotenv
-  - nptyping
-  # interavtive
-  - geoviews
-  - cartopy
-  - ipyleaflet
-  - ipython
-  # viz
-  - seaborn
-  # test
-  - pytest>=6
-  - pytest-cov>=3
-  - deepdiff
-  # docs
-  - sphinx>=4.0
-  - myst-parser
-  - sphinx-copybutton
-  - sphinx-autodoc-typehints
-  - sphinx-book-theme>=0.3
-  - sphinx-design
-  - nbsphinx
-  - nbconvert
-  - pip:
-    - ../.
