+++ env/testing_env.yml
@@ -1,0 +1,8 @@
+name: testing_env
+channels:
+  - conda-forge
+dependencies:
+  - pygmt
+  - pip
+  - pip:
+    - ../.[test]
