### 2024-02-20 aa06cfa
chore: update testing_env and add make command
