### 2024-02-20 6fc1992
fix: delete tmp.nc file in fetch call
