### 2024-02-20 f68def9
build: extend support to python 3.12
