+++ .github/workflows/ci.yml
@@ -51,7 +51,7 @@ jobs:
       # run all of them to catch failures in different combinations.
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
         runs-on: [ubuntu-latest, macos-latest, windows-latest]
 
     env:
