### 2024-02-20 b34d321
chore: list packages after installing local for test GHA
