### 2024-02-20 77c7a0a
chore: reduce sig figs for test_ghf
