+++ tests/test_fetch.py
@@ -1073,8 +1073,8 @@ ghf_test = [
         (
             5000,
             (-3330000.0, 3330000.0, -3330000.0, 3330000.0),
-            26.4593582153,
-            102.412307739,
+            26.5547351837,
+            102.389022827,
             "g",
         ),
     ),
