+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
     "Typing :: Typed",
 ]
 readme = "README.md"
-version = "0.3.0"
+version = "0.3.1"
 license = {file = "LICENSE"}
 
 keywords = ["cryosphere", "antarctica", "arctic", "greenland", "maps", "plotting", "pygmt", "cross-section"]
