+++ .github/workflows/semantic_release.yml
@@ -46,6 +46,8 @@ jobs:
         with:
           fetch-depth: 0
           token: ${{secrets.PAT}}
+      - name: git pull
+        run: git pull origin main --rebase
 
       # use semantic-release
       - id: semantic-release
