### 2024-02-23 #6350 @hugoledoux
@mdtanker thanks for the suggestions of reviewers.

re: the "early development" and your "Ready for daily use but still changing" warning, I think it's known (is it?) that before v1.0 changes can be not backwards compatible.

I'll start working on this week.
