### 2024-02-23 #6350 @mdtanker
@hugoledoux no worries. Yes I think that is the case, let me know if your happy with the updated wording in the readme, or if you'd like me to remove that warning all together.
