### 2024-03-07 58fcf8f
A few typos I spotted while reading the paper
