### 2024-03-07 #6350 @hugoledoux
> @hugoledoux no worries. Yes I think that is the case, let me know if your happy with the updated wording in the readme, or if you'd like me to remove that warning all together.

@mdtanker for now it should be fine, no worries
