### 2024-03-17 #6502 @editorialbot
[REVIEW]: PolarToolkit: Python Tools for Convenient, Reproducible, and Open Polar Science
**Submitting author:** @mdtanker (<a href="http://orcid.org/0000-0003-4266-8554">Matt Tankersley</a>)
**Repository:** https://github.com/mdtanker/polartoolkit
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v0.5.1
