+++ .pre-commit-config.yaml
@@ -26,6 +26,7 @@ exclude: |
       README.md|
       paper/paper.bib|
       paper/paper.md|
+      templates/CHANGELOG.md.j2|
   )$
 
 repos:
