### 2024-03-24 a53f2bb
docs: fix typos pointed out by Hugo
