### 2024-04-23 f6bc045
0.3.2
Automatically generated by python-semantic-release
