### 2024-04-23 4502709
chore: pin python semantic release version
