+++ README.md
@@ -8,40 +8,19 @@
 <!-- SPHINX-START1 -->
 
 <p align="center">
-<a href="https://mybinder.org/v2/gh/mdtanker/polartoolkit/main">
- <img src="https://mybinder.org/badge_logo.svg" alt="Binder link"></a>
- </p>
+<a href="https://mybinder.org/v2/gh/mdtanker/polartoolkit/main"><img src="https://mybinder.org/badge_logo.svg" alt="Binder link"></a>
+</p>
 
 <p align="center">
-<a href="https://pypi.python.org/pypi/polartoolkit">
-<img
-src="https://img.shields.io/pypi/v/polartoolkit?style=flat-square"
-alt="Latest version on PyPI"
-/>
-</a>
-<a href="https://github.com/conda-forge/polartoolkit-feedstock">
-<img
-src="https://img.shields.io/conda/vn/conda-forge/polartoolkit.svg?style=flat-square"
-alt="Latest version on conda-forge"
-/>
-</a>
-<a href="https://codecov.io/gh/mdtanker/polartoolkit">
-<img
-src="https://img.shields.io/codecov/c/github/mdtanker/polartoolkit/main.svg?style=flat-square"
-alt="Test coverage status"
-/>
+<a href="https://pypi.python.org/pypi/polartoolkit"><img src="https://img.shields.io/pypi/v/polartoolkit?style=flat-square"
+alt="Latest version on PyPI"/></a>
+<a href="https://github.com/conda-forge/polartoolkit-feedstock"><img src="https://img.shields.io/conda/vn/conda-forge/polartoolkit.svg?style=flat-square"alt="Latest version on conda-forge"/></a>
+<a href="https://codecov.io/gh/mdtanker/polartoolkit"><img src="https://img.shields.io/codecov/c/github/mdtanker/polartoolkit/main.svg?style=flat-square" alt="Test coverage status"/></a>
+</p>
 
 <p align="center">
-<a href="https://pypi.org/project/polartoolkit/">
-<img src="https://img.shields.io/pypi/pyversions/polartoolkit?style=flat-square"
-alt="Compatible Python versions."
-/>
-</a>
-<a href="https://zenodo.org/badge/latestdoi/475677039">
-<img src="https://zenodo.org/badge/475677039.svg?style=flat-square"
-alt="Zenodo DOI"
-/>
-</a>
+<a href="https://pypi.org/project/polartoolkit/"><img src="https://img.shields.io/pypi/pyversions/polartoolkit?style=flat-square" alt="Compatible Python versions."/></a>
+<a href="https://zenodo.org/badge/latestdoi/475677039"><img src="https://zenodo.org/badge/475677039.svg?style=flat-square" alt="Zenodo DOI"/></a>
 <a href='https://readthedocs.org/projects/polartoolkit/'><img src='https://readthedocs.org/projects/polartoolkit/badge/?version=latest&style=flat-square' alt='Documentation Status' /></a>
  </p>
 
@@ -49,40 +28,34 @@ alt="Zenodo DOI"
 
 ![](docs/cover_fig.png)
 
+**PolarToolkit** (formerly known as Antarctic-Plots) is a Python package with the goal of making Polar (i.e. Antarctic, Arctic, Greenland) research more efficient, reproducible, and accessible. The software does this by providing: 
+
+- Convenient functions for downloading and pre-processing a wide range of commonly used polar datasets
+- Tools for common geospatial tasks (i.e. changing data resolution, subsetting data by geographic regions)
+- Code to easily create publication-quality maps, data profiles, and cross-sections
+
+Additionally, **PolarToolkit** provides an easy means for exploring datasets with pre-defined or interactively-chosen geographic regions.
+
 ## Disclaimer
 
 <p align="center">
[…]
