+++ src/polartoolkit/__init__.py
@@ -9,4 +9,4 @@
 
 from __future__ import annotations
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"
