### 2024-04-23 mdtanker/polartoolkit#193 @JessicaS11
tqdm dependency missing with conda-forge install
**Describe the bug**
Tutorial notebooks require a dev install to run (due to missing dependency `tqdm`).

**Full code that generate the error**

As part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/6502), I followed the [installation instructions for mamba + downloaded the repo using `git clone` here](https://polartoolkit.readthedocs.io/en/latest/install.html). However, I DID NOT perform a dev install (`pip install .`) and attempted to run the `docs/tutorial/fetch_walkthrough.ipynb` notebook. The first call to fetch (`grid = fetch.bedmap2(
    layer="thickness",
    verbose="q",
)` failed with `ValueError: Missing package 'tqdm' required for progress bars.`.

As a user, I would expect that I can follow the installation instructions, clone the repo (or download a tutorial notebook) and it will run without further environment management or local install. I'm not familiar enough with packaging with pyproject to quickly debug, but it looks like somehow `tqdm` isn't getting picked up as a dependency during the conda/mamba installation (is it sill default with `pooch[progress]` as noted in [`pyproject.toml`](https://github.com/mdtanker/polartoolkit/blob/main/pyproject.toml))?).

```python
mamba create --name polartoolkit --yes --force polartoolkit
git clone https://github.com/mdtanker/polartoolkit.git
```
