### 2024-04-23 mdtanker/polartoolkit#194 @JessicaS11
binder not working
**Describe the bug**
The binder environment does not appear to be building. After repeated attempts, it fails while stuck on "Using CondaBuildPack builder".
