### 2024-04-23 #6502 @JessicaS11
Hello @mdtanker! Great toolkit. Below please find my review comments, notwithstanding https://github.com/mdtanker/polartoolkit/issues/193, https://github.com/mdtanker/polartoolkit/issues/194, https://github.com/mdtanker/polartoolkit/issues/62, and agreement with review comments from @pennyhow on adding more information to the documentation.

## Documentation
I really like all the invitations for contributions, including reminders and notes that it's okay if your submission isn't perfect or you don't know how to use all the tools.

> Example Usage

Do you have any examples of a scientific workflow (perhaps as part of a tutorial or publication) containing one of the figures generated by this software?

> Community guidelines

For updates (or new) notebooks, should contributors run them locally and push with outputs, or clear all outputs?

## Software Paper
There's a link typo (to the old repo) on line 38.

> Summary and Statement of Need

The summary is well written, though it suggests the current functionality enables all "Polar" research while in reality it is still limited to Antarctica, with plans to expand to the Arctic/Greenland (yay!). This technicality is important from a user perspective and also because expanding the current package to the north pole will include substantial additions that are not captured in the software release with the current paper.

It would also be helpful for more details in the paper (and documentation) around what types of datasets are available, and what types of polar research they are enabling. For instance, it appears that most (all?) of the source datasets are gridded.

> State of the field

Could you please elaborate on how this fits into the broader tool landscape? For instance, as noted in the documentation, the fetch module uses icepack's EarthDataDownloader class.
