### 2024-04-24 287a4db
build: explicitly specify tqdm as dependency
