### 2024-04-24 mdtanker/polartoolkit#199 @PennyHow
Plot viewing procedures differ in the Download Antarctica datasets tutorial
**Description of the desired feature:** The [Download Antarctica datasets](https://polartoolkit.readthedocs.io/en/latest/tutorial/fetch_walkthrough.html) tutorial does not contain the same plotting procedure for viewing each of the generated figures (compared to the other tutorials). Shouldn't you need a figure priming function after `maps.plot_grd(grid)`? Perhaps something like this:

```python
maps.plot_grd(grid)
fig = maps.plot_grd(grid)
fig.show()
```
