### 2024-04-24 mdtanker/polartoolkit#200 @PennyHow
EarthDataDownloader object to retain user login credentials
**Description of the desired feature:** Functionality added so that login credentials are retained by the `EarthDataDownloader` object after first inputted, and therefore the user does not have to keep re-defining them for each download.

In the tutorials, there is repeated EarthData login requests. For example, in the [3D_stack tutorial](https://polartoolkit.readthedocs.io/en/latest/gallery/3D_stack.html#Add-a-shapefile-to-mask-the-grids,-and-fine-tune-the-layout):

```
measures_shelves = fetch.measures_boundaries(version="IceShelf")
```

Also, there is nothing in your documentation on how to define credentials as environment variables (the icepack link in the [API reference for EarthDataDownloader](https://polartoolkit.readthedocs.io/en/latest/api/polartoolkit.fetch.html#polartoolkit.fetch.EarthDataDownloader) is broken).
