### 2024-04-24 e1efbab
style: auto style fixes from ruff/pylint
