### 2024-04-24 91d3fb3
docs: reuse readme material in RTD index
