### 2024-04-24 #6502 @PennyHow
Thanks for a nice submission, @mdtanker! Generally I feel that this is a great package with good entry-level functionality for beginners in programming, and also handy tools for seasoned cryospheric coders to quickly generate publication-ready figures. The documentation and package structuring reflects these two target user types, with thorough documentation and tutorials.

I have two major comments that need to be addressed before accepting, along with some minor comments:

The functionality is limited to Antarctica
----------------------------------------------------

The term "polar" places an expectation on the package being usable with Antarctic and Arctic datasets. Currently functionality only accomodates datasets for Antarctica. You say that ideally you would like polartoolkit to develop further for use with data from Greenland and the Arctic. On the readthedocs home page, it is stated that this is happening "soon". When is "soon" and would this be within the timeframe of this review process?

Whilst I do not need to see a full implementation of these Greenland/Arctic developments, I think there should be certain changes in the naming conventions and package structuring so that it is easy to foresee how Greenland/Arctic datasets and analysis could be incorporated. For example, I see that the names of the currently available datasets are generically named and not place-specific:

```
>>> import polartoolkit.fetch
>>> print("\n".join(polartoolkit.fetch.get_fetches()))
basal_melt
basement
bedmachine
bedmap2
bedmap_points
crustal_thickness
deepbedmap
etopo
geoid
geomap
ghf
gia
gravity
groundingline
ibcso
ibcso_coverage
ice_vel
imagery
magnetics
mass_change
measures_boundaries
modis_moa
moho
rema
rosetta_gravity
rosetta_magnetics
sediment_thickness
```

These should be changed accordingly so that Greenland/Arctic dataset names could be easily incorporated, i.e. basal_melt >> antarctic_basal_melt.

If you think it is possible, I would also like to request one Greenland/Arctic example be developed and placed in the tutorials/gallery to demonstrate the viability of this future development.

The installation needs to be more straightforward for beginner users
-------------------------------------------------------------------------------------------

The extensive list of package dependencies mean that it takes a long time for mamba/conda to solve the environment with installation. Even the Binder online build did not work for me as it cannot solve the environment in a timely manner.

As this package is partially aimed at beginners in programming, the set-up needs to be simple otherwise it is not accessible for these users. Whilst the installation troubleshooting documentation is good, I think that the package should be modified to make the installation easier. You can solve this problem however you think is best, whether it be changing the package distribution itself, or adding more extensive installation guidance.

Suggestion: I would look at the number of dependencies and refine them by trying to find common functionality between them. And I would also look at alternatives to geopandas, cartopy and pygmt that do not have C bindings (as this is may resolve the issues with the online Binder build).

Minor comments
----------------------

- [ ] https://github.com/mdtanker/polartoolkit/issues/198
- [ ] https://github.com/mdtanker/polartoolkit/issues/199
- [ ] https://github.com/mdtanker/polartoolkit/issues/200

- [ ] The repo is quite big for cloning. Is there anything you can remove to slim down the repo contents and size?
- [ ] For conda installation, please define that users should make sure they should add the conda-forge channel
- [ ] Can you provide some install tests for users in the documentation and repo readme to ensure that their installation works. It could be a simple version call, or some form of unit tests.
- [ ] Please can you add the environment yml file to the repo which you previously provided me. I think it is a good option to have if a user is troubleshooting their install (like me!)

- [ ] Most of the issues with documentation have already been addressed in [this PR](https://github.com/mdtanker/polartoolkit/pull/197)

- [ ] Regarding the state of the field: I think there are other resources/commonly-used packages out there for arctic data handling - check out the [cryosphere awesome list](https://github.com/awesome-cryosphere/cryosphere-links). I would include earthspy at least. I see you also reference [Fatiando](https://www.fatiando.org/) in the readthedocs documentation, which should probably be added here also.

**Other**

- [ ] The logo for polartoolkits looks fairly similar to other software logos I have come across, such as [this](https://github.com/RJArthern/WAVI.jl). It's not really anything that needs specifically changing (maybe it is hard to avoid!), but I just thought I would make you aware that this type of logo is quite common for software related to cryospheric sciences.
