### 2024-04-24 mdtanker/polartoolkit#200 @JessicaS11
If you're not familiar with it, I strongly recommend [earthaccess](https://earthaccess.readthedocs.io/en/latest/). It's a great way to manage Earthdata logins. The library I maintain, [icepyx](https://github.com/icesat2py/icepyx/blob/development/icepyx/core/auth.py) has it set up as an authentication MixIn class so the credentials are accessible from any module that needs them but only asks users for them once.
