### 2024-04-26 mdtanker/polartoolkit#193 @mdtanker
Thanks for opening this. That is strange that `tqdm` isn't being included. I will explicitly add it as a required dependency. Will be fixed as part of #195.
