### 2024-04-26 eeab8f6
chore: cleanup Make commands
