### 2024-04-26 8e2c8ed
docs: simplify binder env
