+++ binder/environment.yml
@@ -1,383 +1,14 @@
-name: polartoolkit
+name: polartoolkit-binder
 channels:
   - conda-forge
+  - nodefaults
 dependencies:
-  - _libgcc_mutex=0.1
-  - _openmp_mutex=4.5
-  - attrs=23.2.0
-  - azure-core-cpp=1.10.3
-  - azure-storage-blobs-cpp=12.10.0
-  - azure-storage-common-cpp=12.5.0
-  - blosc=1.21.5
-  - branca=0.7.0
-  - brotli=1.1.0
-  - brotli-bin=1.1.0
-  - brotli-python=1.1.0
-  - bzip2=1.0.8
-  - c-ares=1.25.0
-  - ca-certificates=2023.11.17
-  - cairo=1.18.0
-  - certifi=2023.11.17
-  - cfitsio=4.3.1
-  - cftime=1.6.3
-  - charset-normalizer=3.3.2
-  - click=8.1.7
-  - click-plugins=1.1.1
-  - cligj=0.7.2
-  - contourpy=1.2.0
-  - curl=8.5.0
-  - cycler=0.12.1
-  - dcw-gmt=2.1.2
-  - expat=2.5.0
-  - fftw=3.3.10
-  - fiona=1.9.5
-  - folium=0.15.1
-  - font-ttf-dejavu-sans-mono=2.37
-  - font-ttf-inconsolata=3.000
-  - font-ttf-source-code-pro=2.038
-  - font-ttf-ubuntu=0.83
-  - fontconfig=2.14.2
-  - fonts-conda-ecosystem=1
-  - fonts-conda-forge=1
-  - fonttools=4.47.2
-  - freetype=2.12.1
-  - freexl=2.0.0
-  - gdal=3.8.3
-  - geopandas=0.14.2
-  - geopandas-base=0.14.2
-  - geos=3.12.1
-  - geotiff=1.7.1
-  - gettext=0.21.1
-  - ghostscript=10.02.1
-  - giflib=5.2.1
-  - gmt=6.5.0
-  - gshhg-gmt=2.3.7
-  - hdf4=4.2.15
-  - hdf5=1.14.3
-  - icu=73.2
-  - idna=3.6
-  - jinja2=3.1.3
-  - joblib=1.3.2
-  - json-c=0.17
-  - kealib=1.5.3
-  - keyutils=1.6.1
-  - kiwisolver=1.4.5
-  - krb5=1.21.2
-  - lcms2=2.16
-  - ld_impl_linux-64=2.40
-  - lerc=4.0.0
-  - libabseil=20230802.1
-  - libaec=1.1.2
-  - libarchive=3.7.2
-  - libblas=3.9.0
-  - libboost-headers=1.84.0
-  - libbrotlicommon=1.1.0
-  - libbrotlidec=1.1.0
-  - libbrotlienc=1.1.0
-  - libcblas=3.9.0
-  - libcrc32c=1.1.2
-  - libcurl=8.5.0
-  - libdeflate=1.19
-  - libedit=3.1.20191231
-  - libev=4.33
-  - libexpat=2.5.0
-  - libffi=3.4.2
-  - libgcc-ng=13.2.0
-  - libgdal=3.8.3
-  - libgfortran-ng=13.2.0
-  - libgfortran5=13.2.0
-  - libglib=2.78.3
-  - libgomp=13.2.0
-  - libgoogle-cloud=2.12.0
-  - libgrpc=1.60.0
-  - libiconv=1.17
-  - libjpeg-turbo=3.0.0
-  - libkml=1.3.0
-  - liblapack=3.9.0
-  - libnetcdf=4.9.2
-  - libnghttp2=1.58.0
[…]
