### 2024-04-26 ccb9b41
refactor: update np random number generator
