### 2024-04-26 53170ed
chore: update ruff version in pre-commit
