+++ .pre-commit-config.yaml
@@ -34,7 +34,7 @@ repos:
         additional_dependencies: [black==23.7.0]
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.1.6"
+    rev: "v0.4.2"
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
