+++ tests/test_utils.py
@@ -8,6 +8,7 @@
 """
 Tests for utils module.
 """
+
 # %%
 from __future__ import annotations
 
