### 2024-04-26 85a5395
docs: remove python version and nodefaults from binder env
