### 2024-04-26 mdtanker/polartoolkit#198 @mdtanker
I tested this and seems the issue is only that there is a blank line between the `if` `elif` and `else` statements. It seems this blank line is no issue in a `.ipynb` or a `.py` but doesn't work in the terminal. Will be fixed with #195
