### 2024-04-26 5ca4de9
docs: try nbgitpull link for launch binder buttons
