### 2024-04-26 671fbef
docs: update RTD links in JOSS paper
