### 2024-04-26 a68049c
docs: switch binder env to separate repo, remove launch buttons
