+++ binder/environment.yml
@@ -1,12 +1,0 @@
-name: polartoolkit-binder
-channels:
-  - conda-forge
-dependencies:
-  - polartoolkit=0.3.3
-  # interactive
-  - geoviews
-  - cartopy
-  - ipyleaflet
-  - ipython
-  # viz
-  - seaborn
