### 2024-04-26 05dbdf2
docs: switch binder env to separate repo, remove launch buttons
