### 2024-04-26 921f091
docs: switch binder env to separate repo, remove launch buttons
