+++ docs/install.md
@@ -4,11 +4,8 @@
 
 See below for the full installation instructions. If instead you'd like to use
 this package online, without needing to install anything, check out our
-[Binder link](https://mybinder.org/v2/gh/mdtanker/polartoolkit/main), which
-gives full access the the package in an online environment.
-
-This Binder environment can also be accessed by clicking the Binder icon in any
-of the `gallery` or `tutorial` examples.
+[Binder link](https://mybinder.org/v2/gh/mdtanker/polartoolkit-binder/main?urlpath=git-pull%3Frepo%3Dhttps%253A%252F%252Fgithub.com%252Fmdtanker%252Fpolartoolkit%26urlpath%3Dtree%252Fpolartoolkit%252Fdocs%252Fgallery%26branch%3Dmain),
+which gives full access the the package in an online environment.
 
 ## Install Python
 
