+++ README.md
@@ -26,8 +26,9 @@ with pre-defined or interactively-chosen geographic regions.
 <!-- SPHINX-START-badges -->
 
 <p align="center">
-<a href="https://mybinder.org/v2/gh/mdtanker/polartoolkit/main"><img src="https://mybinder.org/badge_logo.svg" alt="Binder link"></a>
-</p>
+<a href="https://mybinder.org/v2/gh/mdtanker/polartoolkit-binder/main?urlpath=git-pull%3Frepo%3Dhttps%253A%252F%252Fgithub.com%252Fmdtanker%252Fpolartoolkit%26urlpath%3Dtree%252Fpolartoolkit%252Fdocs%252Fgallery%26branch%3Dmain">
+ <img src="https://mybinder.org/badge_logo.svg" alt="Binder link"></a>
+ </p>
 
 <p align="center">
 <a href="https://pypi.python.org/pypi/polartoolkit"><img src="https://img.shields.io/pypi/v/polartoolkit?style=flat-square"
