### 2024-04-26 mdtanker/polartoolkit#194 @mdtanker
Thanks for checking this. I've tried a few things to fix this, but haven't found a reliable method. For now, until (if) I'm able to reduce PolarToolkit's heavy dependencies, I have moved the Binder environment to a [separate repository](https://github.com/mdtanker/polartoolkit-binder), so that it is not rebuild with every new commit to the main repo.

I've used [nbgitpuller](https://github.com/jupyterhub/nbgitpuller) to create a link which will use the above repo for the build, but the main repo for the content. Below is the badge with this link:

[![Binder](http://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/mdtanker/polartoolkit-binder/main?urlpath=git-pull%3Frepo%3Dhttps%253A%252F%252Fgithub.com%252Fmdtanker%252Fpolartoolkit%26urlpath%3Dtree%252Fpolartoolkit%252Fdocs%252Fgallery%26branch%3Dmain)

Unfortunately, the `sphinx-book-theme` I use for the documentation doesn't allow altering the Binder url, so I've had to disable the `launch-binder` button for all the gallery and user guide examples. Users will just have to navigate to the `ipynb` from the main Binder site. [This issue](https://github.com/executablebooks/sphinx-book-theme/issues/282) should track when we can add back the `launch-binder` buttons.

This issue should close with #195
