### 2024-04-26 mdtanker/polartoolkit#199 @mdtanker
Good catch. It seems Jupyter notebooks will automatically show the figure either way, but to be compatible with running from the terminal, I will add in your correction as part of #195.
