### 2024-04-26 0881b9d
chore: add tests for regions
