### 2024-04-26 cbe0258
style: add typing to function
