+++ docs/datasets/index.md
@@ -1,0 +1,23 @@
+# 📂 Available datasets
+
+Here are all the datasets available currently within PolarToolkit. Please open a
+[Discussion](https://github.com/mdtanker/polartoolkit/discussions) on the GitHub
+page to request new datasets, or check out our
+[Contributor Guide](https://polartoolkit.readthedocs.io/en/latest/contributing.html)
+for learning how to easily contribute a dataset yourself!
+
+## Antarctica
+
+```{nbgallery}
+---
+---
+bedmap2_grids
+```
+
+## Arctic
+
+> No datasets implemented for the Arctic yet.
+
+## Greenland
+
+> No datasets implemented for the Arctic yet.
