### 2024-04-26 ad30f3c
docs: add available dataset gallery
