### 2024-04-28 e3e3540
fix: remove faulty layer name from fetch bedmachine
