+++ docs/_polartoolkit_refs.bib
@@ -49,7 +49,7 @@
 }
 
 @article{baranovantased2021,
-  title = {{{ANTASed}} {\textendash} {{An Updated Sediment Model}} for {{Antarctica}}},
+  title = {{{ANTASed}} -- {{An Updated Sediment Model}} for {{Antarctica}}},
   author = {Baranov, A. and Morelli, A. and Chuvaev, A.},
   year = {2021},
   month = aug,
@@ -59,7 +59,7 @@
   issn = {2296-6463},
   doi = {10.3389/feart.2021.722699},
   urldate = {2021-08-11},
-  abstract = {We compile existing seismic, gravity, radar and magnetic data, together with the subglacial bedrock relief from the BEDMACHINE project, to build the most detailed sediment model for Antarctica. We interpolate these data according to a tectonic map of Antarctica using a statistical kriging method. Our results reveal significant sediment accumulation in Antarctica with several types of sedimentary basins: parts of the Beacon Supergroup and more recent rifting basins. The basement relief closely resembles major geological and tectonic structures. The thickness of sediments has significant variations around the continent, and depends on the degree of crustal extension. West Antarctica has wide sedimentary basins: the Ross basin (thickness 2{\textendash}6 km), the Filchner-Ronne basin (2{\textendash}12 km) with continuations into East Antarctica, the Bentley Subglacial Trench and the Byrd basin (2{\textendash}4 km). The deepest Filchner-Ronne basin has a complex structure with multi-layered sediments. East Antarctica is characterized by vast sedimentary basins such as the Pensacola-Pole (1{\textendash}2 km), Coats Land (1{\textendash}3 km), Dronning Maud Land (1{\textendash}2 km), Vostok (2{\textendash}7 km), Aurora (1{\textendash}3 km), Astrolabe (2{\textendash}4 km), Adventure (2{\textendash}4 km), and Wilkes (1{\textendash}4 km) basins, along with narrow deep rifts filled by sediments: JutulStraumen (1{\textendash}2 km), Lambert (2{\textendash}5 km), Scott, Denman, Vanderford and Totten (2{\textendash}4 km) rifts. The average thickness of sediments for the whole continent is about 0.77 km. The new model, ANTASed, represents a significant improvement over CRUST 1.0 for Antarctica, and reveals new sedimentary basins. Differences between ANTASed and CRUST 1.0 reach +12/-3 km.},
[…]
