### 2024-04-29 mdtanker/polartoolkit#198 @PennyHow
No worries. If Jupiter/pycharm/spyder...etc. is the preferred IDE for running the examples then maybe just add a line at the beginning of the tutorials stating this as a recommendation.
