+++ docs/datasets/index.md
@@ -1,4 +1,4 @@
-# 📂 Available datasets
+# All datasets
 
 Here are all the datasets available currently within PolarToolkit. Please open a
 [Discussion](https://github.com/mdtanker/polartoolkit/discussions) on the GitHub
@@ -6,23 +6,11 @@ page to request new datasets, or check out our
 [Contributor Guide](https://polartoolkit.readthedocs.io/en/latest/contributing.html)
 for learning how to easily contribute a dataset yourself!
 
-## Antarctica
-
-### Topography-related data
-
 ```{nbgallery}
 ---
 ---
-bedmap2_grids
-bedmachine_grids
-basement
-bedmap_points
+antarctica/bedmap2_grids
+antarctica/bedmachine_grids
+antarctica/bedmap_points
+antarctica/basement
 ```
-
-## Arctic
-
-> No datasets implemented for the Arctic yet.
-
-## Greenland
-
-> No datasets implemented for the Arctic yet.
