+++ docs/datasets/greenland/index.md
@@ -1,0 +1,8 @@
+# Greenland
+
+> No datasets added for Greenland yet!
+
+```{nbgallery}
+---
+---
+```
