+++ docs/datasets/arctic/index.md
@@ -1,0 +1,8 @@
+# Arctic
+
+> No datasets added for the Arctic yet!
+
+```{nbgallery}
+---
+---
+```
