### 2024-04-29 88c00ba
docs: reorganize available dataset gallery
