+++ docs/datasets/antarctica/index.md
@@ -1,0 +1,11 @@
+# Antarctica
+
+## Topography-related data
+```{nbgallery}
+---
+---
+bedmap2_grids
+bedmachine_grids
+bedmap_points
+basement
+```
