### 2024-04-29 ca6d0e7
chore: don't run tests on docs or style commits
