### 2024-04-29 8440e6f
chore: don't run tests on docs or style commits
