+++ docs/datasets/antarctica/index.md
@@ -1,6 +1,7 @@
 # Antarctica
 
 ## Topography-related data
+
 ```{nbgallery}
 ---
 ---
