+++ src/polartoolkit/utils.py
@@ -237,13 +237,17 @@ def region_to_df(
 
 
 def region_xy_to_ll(
-    region: tuple[typing.Any, typing.Any, typing.Any, typing.Any], dms: bool = False
+    region: tuple[typing.Any, typing.Any, typing.Any, typing.Any],
+    hemisphere: str,
+    dms: bool = False,
 ) -> tuple[typing.Any, typing.Any, typing.Any, typing.Any]:
     """
     Convert GMT region in format [e, w, n, s] in EPSG:3031 to lat / lon
 
     Parameters
     ----------
+    hemisphere : str,
+        choose between the "north" or "south" hemispheres
     region : tuple[typing.Any, typing.Any, typing.Any, typing.Any]
         region boundaries in GMT format; [e, w, n, s] in meters
     dms: bool, False
@@ -255,7 +259,10 @@ def region_xy_to_ll(
         region boundaries in GMT format; [e, w, n, s] in lat, lon
     """
     df = region_to_df(region)
-    df_proj = epsg3031_to_latlon(df, reg=True)
+    if hemisphere == "north":
+        df_proj = epsg3413_to_latlon(df, reg=True)
+    elif hemisphere == "south":
+        df_proj = epsg3031_to_latlon(df, reg=True)
 
     return tuple([dd2dms(x) for x in df_proj] if dms is True else df_proj)
 
@@ -368,6 +375,110 @@ def epsg3031_to_latlon(
     """
 
     transformer = Transformer.from_crs("epsg:3031", "epsg:4326")
+
+    df_out = df.copy()
+
+    if isinstance(df, pd.DataFrame):
+        (  # pylint: disable=unpacking-non-sequence
+            df_out[output_coord_names[1]],  # type: ignore[call-overload]
+            df_out[output_coord_names[0]],  # type: ignore[call-overload]
+        ) = transformer.transform(
+            df_out[input_coord_names[0]].tolist(),  # type: ignore[call-overload]
+            df_out[input_coord_names[1]].tolist(),  # type: ignore[call-overload]
+        )
+        if reg is True:
+            df_out = [
+                df_out[output_coord_names[0]].min(),  # type: ignore[call-overload]
+                df_out[output_coord_names[0]].max(),  # type: ignore[call-overload]
+                df_out[output_coord_names[1]].min(),  # type: ignore[call-overload]
+                df_out[output_coord_names[1]].max(),  # type: ignore[call-overload]
+            ]
+    else:
+        df_out = list(transformer.transform(df_out[0], df_out[1]))
+    return df_out
+
+
+def latlon_to_epsg3413(
+    df: pd.DataFrame | NDArray[typing.Any, typing.Any],
+    reg: bool = False,
+    input_coord_names: tuple[str, str] = ("lon", "lat"),
+    output_coord_names: tuple[str, str] = ("x", "y"),
+) -> pd.DataFrame | NDArray[typing.Any, typing.Any]:
+    """
+    Convert coordinates from EPSG:4326 WGS84 in decimal degrees to EPSG:3413 North Polar
+    Stereographic in meters.
+
+    Parameters
+    ----------
+    df : pd.DataFrame or NDArray[typing.Any, typing.Any]
+        input dataframe with latitude and longitude columns
+    reg : bool, optional
+        if true, returns a GMT formatted region string, by default False
+    input_coord_names : list, optional
+        set names for input coordinate columns, by default ["lon", "lat"]
+    output_coord_names : list, optional
+        set names for output coordinate columns, by default ["x", "y"]
+
+    Returns
+    -------
+    pd.DataFrame or NDArray[typing.Any, typing.Any]
+        Updated dataframe with new easting and northing columns or NDArray in format
+        [e, w, n, s]
[…]
