### 2024-04-30 6de294d
feat: add pen kwarg to show_region
