### 2024-04-30 c8aa365
fix: load deepbedmap instead of returning string
