### 2024-04-30 ac873db
docs: edit some docstrings
