### 2024-04-30 b6ecc54
docs: edit some docstrings
