+++ docs/datasets/index.md
@@ -9,8 +9,19 @@ for learning how to easily contribute a dataset yourself!
 ```{nbgallery}
 ---
 ---
-antarctica/bedmap2_grids
-antarctica/bedmachine_grids
+antarctica/bedmap2
+antarctica/bedmachine
+antarctica/deepbedmap
+antarctica/etopo
+antarctica/rema
+antarctica/ibcso
+antarctica/geoid
 antarctica/bedmap_points
+antarctica/ibcso_coverage
 antarctica/basement
+antarctica/moho
+antarctica/crustal_thickness
+antarctica/sediment_thickness
+antarctica/imagery
+antarctica/modis_moa
 ```
