+++ .pre-commit-config.yaml
@@ -43,6 +43,7 @@ repos:
     rev: "v4.5.0"
     hooks:
       - id: check-added-large-files
+        exclude: ^docs/datasets/
       - id: check-case-conflict
       - id: check-merge-conflict
       - id: check-symlinks
