### 2024-05-08 fbe97f3
docs: add more dataset notebooks
