+++ docs/datasets/antarctica/index.md
@@ -5,8 +5,26 @@
 ```{nbgallery}
 ---
 ---
-bedmap2_grids
-bedmachine_grids
+bedmap2
+bedmachine
+deepbedmap
+etopo
+rema
+ibcso
+geoid
 bedmap_points
+ibcso_coverage
 basement
+moho
+crustal_thickness
+sediment_thickness
+```
+
+## Imagery
+
+```{nbgallery}
+---
+---
+imagery
+modis_moa
 ```
