+++ docs/datasets/greenland/index.md
@@ -5,4 +5,5 @@
 ```{nbgallery}
 ---
 ---
+modis_mog
 ```
