### 2024-05-10 39fca9b
feat: add subset grid function
