### 2024-05-10 e532d6a
feat: add subset grid function
