### 2024-05-10 ab2e60a
feat: add subset grid function
