### 2024-05-10 e2b214d
feat: add hemisphere arg to most plotting functions
BREAKING CHANGE: this alters a vast majority of the code!
