+++ src/polartoolkit/maps.py
@@ -55,6 +55,7 @@ except ImportError:
 
 def basemap(
     region: tuple[float, float, float, float] | None = None,
+    hemisphere: str | None = None,
     fig_height: float = 15,
     fig_width: float | None = None,
     origin_shift: str = "initialize",
@@ -67,6 +68,9 @@ def basemap(
     ----------
     region : tuple[float, float, float, float] | None, optional
         bounding region in GMT format, by default None
+    hemisphere : str, optional
+        set whether to plot in "north" hemisphere (EPSG:3413) or "south" hemisphere
+        (EPSG:3031), only used if plot lat long components (gridlines)
     fig_height : float, optional
         height of figure, by default 15
     fig_width : float | None, optional
@@ -89,12 +93,14 @@ def basemap(
     if fig_width is None:
         proj, proj_latlon, fig_width, fig_height = utils.set_proj(
             region,
+            hemisphere=hemisphere,
             fig_height=fig_height,
         )
     # if fig_width is set, use it to set projection
     else:
         proj, proj_latlon, fig_width, fig_height = utils.set_proj(
             region,
+            hemisphere=hemisphere,
             fig_width=fig_width,
         )
 
@@ -124,19 +130,25 @@ def basemap(
         verbose="e",
     )
 
-    # plot coast
+    # plot groundingline and coastlines
     if kwargs.get("coast", False) is True:
         add_coast(
             fig,
-            region,
-            proj,
+            hemisphere=hemisphere,
+            region=region,
+            projection=proj,
             pen=kwargs.get("coast_pen", None),
             no_coast=kwargs.get("no_coast", False),
-            version=kwargs.get("coast_version", "depoorter-2013"),
+            version=kwargs.get("coast_version", None),
         )
 
     # add lat long grid lines
     if kwargs.get("gridlines", False) is True:
+        if hemisphere is None:
+            logging.warning(
+                "Argument `hemisphere` not specified, will use meters for gridlines."
+            )
+
         add_gridlines(
             fig,
             region=region,
@@ -158,12 +170,16 @@ def basemap(
         }
         add_inset(
             fig,
-            # inset_pos=kwargs.get("inset_pos", "TL"),
+            hemisphere=hemisphere,  # type: ignore[arg-type]
             **new_kwargs,
         )
 
     # add scalebar
     if kwargs.get("scalebar", False) is True:
+        if proj_latlon is None:
+            msg = "Argument `hemisphere` needs to be specified for plotting a scalebar"
+            raise ValueError(msg)
+
         add_scalebar(
             fig=fig,
             region=region,
@@ -194,6 +210,7 @@ def basemap(
 
 def plot_grd(
     grid: str | xr.DataArray,
+    hemisphere: str | None = None,
     cmap: str | bool = "viridis",
     region: tuple[float, float, float, float] | None = None,
     coast: bool = False,
@@ -201,19 +218,22 @@ def plot_grd(
     fig: pygmt.Figure | None = None,
     **kwargs: typing.Any,
 ) -> pygmt.Figure:
-    r"""
+    """
     Helps easily create PyGMT maps, individually or as subplots.
 
     Parameters
     ----------
     grid : str or xr.DataArray
         grid file to plot, either loaded xr.DataArray or string of a filename
+    hemisphere : str, optional
+        set whether to plot in "north" hemisphere (EPSG:3413) or "south" hemisphere
+        (EPSG:3031), only used if plot lat long components (gridlines)
     cmap : str or bool, optional
         GMT color scale to use, by default 'viridis'
     region : tuple[float, float, float, float], optional
         region to plot, by default is extent of input grid
     coast : bool, optional
[…]
