### 2024-05-10 3a03996
feat: add hemisphere arg to most plotting functions
BREAKING CHANGE: this alters a vast majority of the code!
