### 2024-05-10 05cb358
feat: add hemisphere arg to most plotting functions
BREAKING CHANGE: this alters a vast majority of the code!
