### 2024-05-10 677b13e
chore: change arg `image` to `modis`
