### 2024-05-10 dabeb44
chore: add greenland coast version to plotting functions
