### 2024-05-10 f0f5887
fix: correctly center interactive map
