### 2024-05-10 e633789
feat: allow passing .cpt file to colorbar histogram
