### 2024-05-10 6b05fb9
fix: improve creation of colormap
