### 2024-05-10 ee155de
feat: add function for combined min max of several grids
