### 2024-05-10 681e95a
docs: add modis mog to index
