+++ tests/test_utils.py
@@ -154,19 +154,19 @@ def test_region_xy_to_ll_north():
     reg_ll = utils.region_xy_to_ll(reg_xy, hemisphere="north", dms=True)
 
     assert reg_ll == (
-        "-92:17:26.19615349866217",
-        "11:18:35.75690647277224",
-        "58:19:54.14523478713818",
-        "81:50:50.37008428043919",
+        "-82:34:6.931303778954316",
+        "-2:17:26.19615349869855",
+        "77:39:39.3112071675132",
+        "82:26:25.183721040550154",
     )
 
     reg_ll = utils.region_xy_to_ll(reg_xy, hemisphere="north")
 
     assert reg_ll == (
-        -92.29061004263852,
-        11.309932474020215,
-        58.33170700966309,
-        81.84732502341123,
+        -82.56859202882748,
+        -2.2906100426385274,
+        77.66091977976876,
+        82.44032881140015,
     )
 
 
