### 2024-05-13 #6502 @mdtanker
Hi all, thank you all for the prompt and thorough reviews! I really appreciate all the time you have put in. Apologies for not responding sooner, but I've been trying to address many of the suggestions before doing a line-by-line response. I've made lots of commits in the past two weeks to the [PR: Edits for JOSS reviews](https://github.com/mdtanker/polartoolkit/pull/195).

Some of these suggestions will take some more time to correctly implement, but I anticipate being done within 2-3 weeks. The major things left for me to do are add all the available datasets to a gallery, add a few more Greenland-specific datasets, and try and reduce the dependencies and size of the repo.

Once I have accomplished most of these tasks, I will address all your comments here :smiley:
