### 2024-06-03 d999486
refactor: change default point  color in plot_profile
