### 2024-06-03 830c9bf
fix: warning about histogram if grid is constant value
