### 2024-06-03 b71001a
feat: enable passing single grid to plot_3d
