+++ src/polartoolkit/maps.py
@@ -208,192 +208,55 @@ def basemap(
     return fig
 
 
-def plot_grd(
-    grid: str | xr.DataArray,
+def set_cmap(
+    cmap: str | bool,
+    grid: str | xr.DataArray | None = None,
+    modis: bool = False,
+    grd2cpt: bool = False,
+    cpt_lims: tuple[float, float] | None = None,
+    cmap_region: tuple[float, float, float, float] | None = None,
+    robust: bool = False,
+    reverse_cpt: bool = False,
+    shp_mask: gpd.GeoDataFrame | None = None,
     hemisphere: str | None = None,
-    cmap: str | bool = "viridis",
-    region: tuple[float, float, float, float] | None = None,
-    coast: bool = False,
-    origin_shift: str = "initialize",
-    fig: pygmt.Figure | None = None,
+    colorbar: bool = True,
     **kwargs: typing.Any,
-) -> pygmt.Figure:
+) -> tuple[str | bool, bool, tuple[float, float] | None]:
     """
-    Helps easily create PyGMT maps, individually or as subplots.
+    Function used to set the PyGMT colormap for a figure.
 
     Parameters
     ----------
-    grid : str or xr.DataArray
-        grid file to plot, either loaded xr.DataArray or string of a filename
-    hemisphere : str, optional
-        set whether to plot in "north" hemisphere (EPSG:3413) or "south" hemisphere
-        (EPSG:3031), only used if plot lat long components (gridlines)
-    cmap : str or bool, optional
-        GMT color scale to use, by default 'viridis'
-    region : tuple[float, float, float, float], optional
-        region to plot, by default is extent of input grid
-    coast : bool, optional
-        choose whether to plot coastline and grounding line, by default False
-    origin_shift : str, optional
-        automatically will create a new figure, set to 'xshift' to instead add plot to
-        right of previous plot, or 'yshift' to add plot above previous plot, by
-        default 'initialize'.
-    fig : pygmt.Figure(), optional
-        supply figure instance for adding subplots or other PyGMT plotting methods, by
-        default None
-
-    Keyword Args
-    ------------
-    modis : bool
-        set to True if plotting MODIS data to use a nice colorscale.
-    grd2cpt : bool
-        use GMT module grd2cpt to set color scale from grid values, by default is False
-    cmap_region : str or tuple[float, float, float, float]
-        region to use to define color scale if grd2cpt is True, by default is
-        region
-    cbar_label : str
-        label to add to colorbar.
-    points : pd.DataFrame
-        points to plot on map, must contain columns 'x' and 'y'.
-    show_region : tuple[float, float, float, float]
-        GMT-format region to use to plot a bounding regions.
-    cpt_lims : str or tuple]
-        limits to use for color scale max and min, by default is max and min of data.
-    gridlines : bool
-        choose to plot lat/long grid lines, by default is False
-    inset : bool
-        choose to plot inset map showing figure location, by default is False
-    inset_pos : str
-        position for inset map; either 'TL', 'TR', BL', 'BR', by default is 'TL'
-    fig_height : int or float
-        height in cm for figures, by default is 15cm.
-    scalebar: bool
-        choose to add a scalebar to the plot, by default is False. See
-        `maps.add_scalebar` for additional kwargs.
-    colorbar: bool
-        choose to add a colorbar to the plot, by default is True
+    cmap : str | bool
+        a string of either a PyGMT cpt file (.cpt), or a preset PyGMT color ramp, or
+        alternatively a value of True will use the last used cmap.
+    grid : str | xr.DataArray | None, optional
+       grid used for grd2cpt colormap equalization, by default None
+    modis : bool, optional
[…]
