### 2024-06-03 f7395b6
fix: update figure shifting of plot_3d
