### 2024-06-03 7271017
fix: update warning about grid region extract in plot_3d
