### 2024-06-03 adc5168
feat: add cbar_perspective arg to plot_3d
