### 2024-06-03 49cd2f2
refactor: remove default vlims for plot_3d
