### 2024-06-03 9d26d06
feat: add cbar_labels kwarg to plot_3d
