### 2024-06-03 f7d0083
refactor: changed default shading to false in plot_3d
