### 2024-06-03 7100cb8
refactor: set cbar position in plot_3d
