### 2024-06-03 1e68076
refactor: use new set_cmap function in plot_3d
