### 2024-06-03 6286ccd
docs: add GHF to dataset gallery
