### 2024-06-03 95826c1
docs: add GHF to dataset gallery
