### 2024-06-03 2200e0b
docs: add GIA to dataset gallery
