+++ docs/datasets/index.md
@@ -25,5 +25,6 @@ antarctica/sediment_thickness
 antarctica/imagery
 antarctica/modis_moa
 antarctica/geothermal_heat_flux
+antarctica/glacial_isostatic_adjustment
 greenland/modis_mog
 ```
