### 2024-06-03 c25c592
docs: add GIA to dataset gallery
