+++ docs/datasets/antarctica/index.md
@@ -35,4 +35,5 @@ modis_moa
 ---
 ---
 geothermal_heat_flux
+glacial_isostatic_adjustment
 ```
