### 2024-06-03 c470dee
docs: add ice velocity to dataset gallery
