### 2024-06-03 3e8a022
docs: add ice velocity to dataset gallery
