### 2024-06-03 c9a222a
docs: add magnetics to dataset gallery
