+++ docs/datasets/index.md
@@ -27,6 +27,7 @@ antarctica/modis_moa
 antarctica/geothermal_heat_flux
 antarctica/glacial_isostatic_adjustment
 antarctica/gravity
+antarctica/magnetics
 antarctica/ice_velocity
 greenland/modis_mog
 ```
