### 2024-06-03 f0cadd9
docs: add magnetics to dataset gallery
