+++ docs/datasets/antarctica/index.md
@@ -46,4 +46,5 @@ magnetics
 geothermal_heat_flux
 glacial_isostatic_adjustment
 ice_velocity
+mass_change
 ```
