### 2024-06-03 39b142e
docs: add groundline to dataset gallery
