### 2024-06-03 48f9e8b
docs: add groundline to dataset gallery
