### 2024-06-03 27dfe4d
docs: add groundline to dataset gallery
