### 2024-06-03 c328e3e
docs: add geomap geology to dataset gallery
