+++ docs/datasets/antarctica/index.md
@@ -55,4 +55,5 @@ mass_change
 ---
 ---
 groundingline
+geology
 ```
