### 2024-06-03 4277b7c
docs: add geomap geology to dataset gallery
