+++ docs/datasets/greenland/index.md
@@ -1,6 +1,5 @@
 # Greenland
 
-
 ## Imagery
 
 ```{nbgallery}
