### 2024-06-03 c50860d
docs: add basal melt rate to dataset gallery
