### 2024-06-03 e6827db
docs: add basal melt rate to dataset gallery
