+++ docs/datasets/antarctica/index.md
@@ -47,6 +47,7 @@ geothermal_heat_flux
 glacial_isostatic_adjustment
 ice_velocity
 mass_change
+basal_melt
 ```
 
 ## Shapefiles and Misc.
