### 2024-06-04 f1cd037
docs: add measures ice boundaries to dataset gallery
