+++ docs/datasets/antarctica/index.md
@@ -56,5 +56,6 @@ basal_melt
 ---
 ---
 groundingline
+ice_boundaries
 geology
 ```
