### 2024-06-04 c36d034
docs: add measures ice boundaries to dataset gallery
