### 2024-06-04 bf9c1cd
style: fixes for pylint
