### 2024-06-04 76b9865
docs: add rosetta grav/mag to gallery datasets
