### 2024-06-04 42d31eb
docs: add rosetta grav/mag to gallery datasets
