### 2024-06-05 d5aeff8
chore: pre-commit autoupdate to monthly
