### 2024-06-05 a652a3f
docs: clarify project scope and goals
