### 2024-06-05 69450ac
docs: clarify all mentions of format of region tuples
