### 2024-06-05 1e8c09f
docs: clarify all mentions of format of region tuples
