+++ src/polartoolkit/fetch.py
@@ -2859,217 +2859,6 @@ def geoid(
     return typing.cast(xr.DataArray, resampled)
 
 
-def rosetta_gravity(version: str = "gravity") -> pd.DataFrame:
-    """
-    Load either a shapefile of ROSETTA-ice flightlines, a dataframe of ROSETTA-Ice
-    airborne gravity data over the Ross Ice Shelf, or a dataframe of ROSETTA-Ice density
-    values from the density inversion.
-    from :footcite:t:`tintoross2019`.
-    Accessed from https://www.usap-dc.org/view/project/p0010035
-    This is only data from the first 2 of the 3 field seasons.
-    Columns:
-    Line Number: The ROSETTA-Ice survey line number, >1000 are tie lines
-    Latitude (degrees): Latitude decimal degrees WGS84
-    Longitude (degrees): Longitude decimal degrees WGS84
-    unixtime (seconds): The number of seconds that have elapsed since midnight
-    (00:00:00 UTC) on January 1st, 1970
-    Height (meters): Height above WGS84 ellipsoid
-    x (meters): Polar stereographic projected coordinates true to scale at 71° S
-    y (meters): Polar stereographic projected coordinates true to scale at 71° S
-    FAG_levelled (mGal): Levelled free air gravity
-
-    Parameters
-    ----------
-    version : str, optional
-
-    Returns
-    -------
-    pd.DataFrame
-        Returns a dataframe containing the gravity, density, or flightline data
-
-    References
-    ----------
-    .. footbibliography::
-    """
-
-    if version == "shapefile":
-        path = pooch.retrieve(
-            url="http://wonder.ldeo.columbia.edu/data/ROSETTA-Ice/GridInformation/Shapefile/ROSETTA-Ice_Grid_Flown_Shapefile.zip",
-            fname="ROSETTA-Ice_Grid_Flown_Shapefile.zip",
-            path=f"{pooch.os_cache('pooch')}/polartoolkit/gravity",
-            known_hash="17e0026fe386503d74c679a6091235e2394fb269030f5555372e55bdf61f6946",
-            progressbar=True,
-            processor=pooch.Unzip(),
-        )
-        # path to shapefile
-        fname = next(p for p in path if p.endswith(".shp"))
-
-        # read the file into a geodataframe
-        df = pyogrio.read_dataframe(fname)
-    elif version == "gravity":
-        path = pooch.retrieve(
-            url="http://wonder.ldeo.columbia.edu/data/ROSETTA-Ice/Gravity/rs_2019_grav.csv",
-            fname="ROSETTA_2019_grav.csv",
-            path=f"{pooch.os_cache('pooch')}/polartoolkit/gravity",
-            known_hash="d4fdfcc293ac13e6222938bfa9e1e9ccc781ea7556a13d356e9f1b4aba809928",
-            progressbar=True,
-        )
-
-        df = pd.read_csv(path)
-
-        # convert line numbers into float format (L200 -> 200)
-        df.Line = df.Line.str[1:]
-        df["Line"] = pd.to_numeric(df["Line"])
-
-    elif version == "density":
-        path = pooch.retrieve(
-            url="http://wonder.ldeo.columbia.edu/data/ROSETTA-Ice/DerivedProducts/Density/rs_2019_density.csv",
-            fname="rs_2019_density.csv",
-            path=f"{pooch.os_cache('pooch')}/polartoolkit/gravity",
-            known_hash="8c3d6f69087095ab1a184d0517a136ebd59e4761110601b30da1ef3685883617",
-            progressbar=True,
-        )
-
-        df = pd.read_csv(path)
-
[…]
