### 2024-06-05 cc8d2af
fix: remove support for fetch ROSETTA grav and mag to align with goal of only providing common and widespread datasets
