+++ docs/datasets/antarctica/index.md
@@ -36,8 +36,6 @@ modis_moa
 ---
 gravity
 magnetics
-rosetta_gravity
-rosetta_magnetics
 ```
 
 ## Geophysically-derived data
