### 2024-06-05 eb90b15
fix: remove fetch basement due to limited spatial extent of data
