### 2024-06-05 86be2e5
fix: remove fetch basement due to limited spatial extent of data
