### 2024-06-05 4cfc9e5
fix: remove fetch basement due to limited spatial extent of data
