+++ src/polartoolkit/fetch.py
@@ -1066,75 +1066,6 @@ def measures_boundaries(
     return fname
 
 
-def basement(
-    region: tuple[float, float, float, float] | None = None,
-    spacing: float | None = None,
-    registration: str | None = None,
-) -> xr.DataArray:
-    """
-    Load a grid of basement topography.
-    Offshore and sub-Ross Ice Shelf basement topography.
-    from :footcite:t:`tankersleybasement2022`, :footcite:t:`tankersleybasement2022a` and
-    :footcite:t:`lindequepreglacial2016a`.
-
-    Elevations are referenced to WGS84 ellipsoid.
-
-    Parameters
-    ----------
-    region : tuple[float, float, float, float], optional
-        region to clip the loaded grid to, in format [xmin, xmax, ymin, ymax], by
-        default doesn't clip
-    spacing : float, optional
-        grid spacing to resample the loaded grid to, by default 5e3
-    registration : str, optional
-        change registration with either 'p' for pixel or 'g' for gridline registration,
-        by default is None.
-
-    Returns
-    -------
-    xr.DataArray
-        dataarray of basement depths
-
-    References
-    ----------
-    .. footbibliography::
-    """
-
-    # found with utils.get_grid_info()
-    initial_region = (-3330000.0, 1900000.0, -3330000.0, 1850000.0)
-    initial_spacing = 5e3
-    initial_registration = "p"
-
-    if region is None:
-        region = initial_region
-    if spacing is None:
-        spacing = initial_spacing
-    if registration is None:
-        registration = initial_registration
-
-    path = pooch.retrieve(
-        url="https://download.pangaea.de/dataset/941238/files/Ross_Embayment_basement_filt.nc",
-        fname="basement.nc",
-        path=f"{pooch.os_cache('pooch')}/polartoolkit/basement",
-        known_hash="2b4fd53d1bb766d265ba7d504bf446fe4e4c77a5f4ab568892908edf385d9c67",
-        progressbar=True,
-    )
-
-    grid = xr.load_dataarray(path)
-
-    resampled = resample_grid(
-        grid,
-        initial_spacing=initial_spacing,
-        initial_region=initial_region,
-        initial_registration=initial_registration,
-        spacing=spacing,
-        region=region,
-        registration=registration,
-    )
-
-    return typing.cast(xr.DataArray, resampled)
-
-
 def sediment_thickness(
     version: str,
     region: tuple[float, float, float, float] | None = None,
