### 2024-06-05 a86e50e
refactor: rename fetch `measures_boundaries` to `antarctic_boundaries`
BREAKING CHANGE: make sure to update your code with the new function name!
