### 2024-06-10 cfa5fff
feat: add bedmachine data from Greenland
