+++ docs/datasets/greenland/index.md
@@ -6,6 +6,7 @@
 ---
 ---
 bedmachine
+geoid
 ```
 
 ## Imagery
