### 2024-06-10 914c88f
fix: for bedmachine grids, restore correct registration type and only resample after geoid added to grids
