### 2024-06-10 3c04935
chore: update tests for bedmachine with reference change
