### 2024-06-10 b8b4f0a
feat: add function to extract default hemisphere
