### 2024-06-10 d756838
refactor: if available, use environmental variables for setting hemisphere throughout code
To utilize this feature, either set a system environment variable POLARTOOLKIT_HEMISPHERE equal to either south or north, or in your python file/notebook use os.environ to set this value for the current session.
