### 2024-06-10 3be6d86
refactor: use environment variable for setting hemisphere in profile
