### 2024-06-10 9fa6c73
refactor: log info instead of error in mask_from_shp
