### 2024-06-10 8783c6d
feat: enable greenland support for profiles
