### 2024-06-11 81245ba
chore: update use of pylint
