+++ noxfile.py
@@ -40,7 +40,7 @@ def pylint(session: nox.Session) -> None:
     """
     # This needs to be installed into the package environment, and is slower
     # than a pre-commit check
-    session.install(".", "pylint")
+    session.install(".", "pylint>=3.2")
     session.run("pylint", "polartoolkit", *session.posargs)
 
 
