### 2024-06-11 a93d83a
build: set min pylint version
