### 2024-06-11 fdab8fe
refactor: remove unnecessary verbose and kwargs for some fetches
