### 2024-06-11 b0a9805
style: fix cmap issues in plot_profile
