### 2024-06-11 7c47109
refactor: add hemisphere and remove kwargs from call to geoid
