### 2024-06-11 6e7ea93
fix: specify hemisphere in bedmap2 reference test
