### 2024-06-11 mdtanker/polartoolkit#209 @mdtanker
New Docs structure
I think we should update the documentation based on the [Divio system](https://documentation.divio.com/).  This will switch from our current style of a `gallery` and `tutorials` to a five category system consisting of: `Getting started`, `Tutorial`,  `How-to`, `Reference`, and `Explanation`.  Inspired by [this Verde issue](https://github.com/fatiando/verde/issues/433)

Simple cooking analogy to understand these sections:
* `Tutorial` is a cooking lesson
* `How-to` is a recipe
* `Reference` is a encyclopedia article about an ingredient
* `Explanation` is a book on the history and theory of cooking

# Getting started
- [ ] Installing
- [ ] Notebook with a quick taste of PolarToolkit. choose a pre-set region (Ross Ice Shelf), download data (Bedmap2), make a map, add standard embellishments, and plot a star to point out Disco Deep, and a cross-section through the region
- [ ] Citing PolarToolkit

# Tutorial
These tutorials should grow in complexity and are meant to be followed in order by someone new to PolarToolkit. They shouldn't try to explain how things work, but just teach by doing.
#### Loading data (Fetch)
- [ ] Download Polar data: Download 1 dataset each from Antarctica and Greenland with defaults. Not plotting
- [ ] Specify grid parameters: choose region, spacing, and reference frame
#### Creating plots (Maps)
- [ ] Create a simple map: Download 1 dataset and create a simple map with defaults
- [ ] Add map elements: add colorbar, gridlines, inset, north arrow, scalebar
- [ ] Make subplots: Download 4 datasets and make a subplot
- [ ] Show gridded data in 3D: Download bed topography data and plot in 3D
- [ ] Plot point data: Download some point data and use maps.basemap and PyGMT to plot it
#### Creating profiles (Profile)
- [ ] Create a cross-section: default profile with 2 endpoints
- [ ] Show data along a profile: add default data
#### Other
- [ ] Interactively explore your data: interactive gridded and point data
- [ ] Reproject your data: load a grid and some point data in lat long, reproject to EPSG.
- [ ]

# How-to
These should be short and to the point, focused on the problem presented. They can assume that people have done the tutorials. Also don't explain things, but allow some flexibility.
- [ ] Download NASA data
- [ ] Switch between Antarctica, the Arctic or Greenland
- [ ] Define a geographic region: manually (EPSG grid figures), bounding data, or interactively
- [ ] Manipulating geographic regions: combining, expanding
- [x] Use with PyGMT: define projection and use normal PyGMT plotting functions
- [x] Choose a colorscale
- [x] Organize subplots
- [ ] Grid manipulations: subset, resample, filter, change registration
- [ ] Point data manipulations: block reduce, inside region,
- [ ] Compare grids
- [ ] Mask data: with shapefile and interactively
- [ ] Customize profiles: clip ends, change basemap style, etc.
- [ ] Extract grid info: get_grd_info, get min max, get combined min max

# Reference
- [ ] Available regions
- [ ] Available datasets
- [ ] API
- [ ] References
- [ ] Changelog

# Explanation
These are meant to explain how things work and why they are that way.
- [ ] EPSG projections
- [ ] Specifying which hemisphere
- [ ]
