### 2024-06-11 e4acbeb
chore: exclude some md files from pre-commit
