+++ .pre-commit-config.yaml
@@ -6,6 +6,10 @@ ci:
 exclude: |
   (?x)^(
       docs/index.md|
+      docs/install.md|
+      docs/citing.md|
+      docs/overview.md|
+      docs/reference.md|
       docs/api/polartoolkit.rst|
       docs/api/polartoolkit.fetch.rst|
       docs/api/polartoolkit.maps.rst|
@@ -19,6 +23,7 @@ exclude: |
       .github/config.yml|
       CONTRIBUTING.md|
       templates/CHANGELOG.md.j2|
+      README.md|
   )$
 
 repos:
