### 2024-06-11 8ef681f
build: include environment.yml file as backup for issues with installing
