+++ env/environment.yml
@@ -1,0 +1,270 @@
+name: polartoolkit
+channels:
+  - conda-forge
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_gnu
+  - affine=2.4.0=pyhd8ed1ab_0
+  - asciitree=0.3.3=py_2
+  - attrs=23.2.0=pyh71513ae_0
+  - aws-c-auth=0.7.22=h96bc93b_2
+  - aws-c-cal=0.6.14=h88a6e22_1
+  - aws-c-common=0.9.19=h4ab18f5_0
+  - aws-c-compression=0.2.18=h83b837d_6
+  - aws-c-event-stream=0.4.2=ha47c788_12
+  - aws-c-http=0.8.1=h29d6fba_17
+  - aws-c-io=0.14.8=h21d4f22_5
+  - aws-c-mqtt=0.10.4=h759edc4_4
+  - aws-c-s3=0.5.9=h594631b_3
+  - aws-c-sdkutils=0.1.16=h83b837d_2
+  - aws-checksums=0.1.18=h83b837d_6
+  - aws-crt-cpp=0.26.9=he3a8b3b_0
+  - aws-sdk-cpp=1.11.329=hba8bd5f_3
+  - azure-core-cpp=1.11.1=h91d86a7_1
+  - azure-identity-cpp=1.6.0=hf1915f5_1
+  - azure-storage-blobs-cpp=12.10.0=h00ab1b0_1
+  - azure-storage-common-cpp=12.5.0=h94269e2_4
+  - blosc=1.21.5=hc2324a3_1
+  - bokeh=3.4.1=pyhd8ed1ab_0
+  - branca=0.7.2=pyhd8ed1ab_0
+  - brotli=1.1.0=hd590300_1
+  - brotli-bin=1.1.0=hd590300_1
+  - brotli-python=1.1.0=py312h30efb56_1
+  - bzip2=1.0.8=hd590300_5
+  - c-ares=1.28.1=hd590300_0
+  - ca-certificates=2024.6.2=hbcca054_0
+  - cairo=1.18.0=h3faef2a_0
+  - certifi=2024.6.2=pyhd8ed1ab_0
+  - cfitsio=4.4.0=hbdc6101_1
+  - cftime=1.6.3=py312h085067d_1
+  - charset-normalizer=3.3.2=pyhd8ed1ab_0
+  - click=8.1.7=unix_pyh707e725_0
+  - click-plugins=1.1.1=py_0
+  - cligj=0.7.2=pyhd8ed1ab_1
+  - cloudpickle=3.0.0=pyhd8ed1ab_0
+  - contourpy=1.2.1=py312h8572e83_0
+  - curl=8.8.0=he654da7_0
+  - cycler=0.12.1=pyhd8ed1ab_0
+  - cytoolz=0.12.3=py312h98912ed_0
+  - dask=2024.5.2=pyhd8ed1ab_0
+  - dask-core=2024.5.2=pyhd8ed1ab_0
+  - dask-expr=1.1.2=pyhd8ed1ab_0
+  - dcw-gmt=2.1.2=ha770c72_0
+  - distributed=2024.5.2=pyhd8ed1ab_0
+  - docrep=0.3.2=pyh44b312d_0
+  - et_xmlfile=1.1.0=pyhd8ed1ab_0
+  - expat=2.6.2=h59595ed_0
+  - fasteners=0.17.3=pyhd8ed1ab_0
+  - fftw=3.3.10=nompi_hf1063bd_110
+  - fiona=1.9.6=py312h32ad294_3
+  - fmt=10.2.1=h00ab1b0_0
[…]
