### 2024-06-11 f1892b8
docs: consolidate and update install instructions
