+++ README.md
@@ -7,19 +7,14 @@
 
 <!-- SPHINX-START-proj-desc -->
 
-**PolarToolkit** (formerly known as Antarctic-Plots) is a Python package with
-the goal of making Polar (i.e. Antarctic, Arctic, Greenland) research more
-efficient, reproducible, and accessible. The software does this by providing:
+**PolarToolkit** (formerly known as Antarctic-Plots) is a Python package to make Polar (i.e. Antarctic, Arctic, Greenland) research more efficient, reproducible, and accessible.
+The software does this by providing:
 
-- Convenient functions for downloading and pre-processing a wide range of
-  commonly used polar datasets
-- Tools for common geospatial tasks (i.e. changing data resolution, subsetting
-  data by geographic regions)
-- Code to easily create publication-quality maps, data profiles, and
-  cross-sections
+- Convenient functions for downloading and pre-processing a wide range of commonly used polar datasets
+- Tools for common geospatial tasks (i.e. changing data resolution, subsetting data by geographic regions)
+- Code to easily create publication-quality maps, data profiles, and cross-sections
 
-Additionally, **PolarToolkit** provides an easy means for exploring datasets
-with pre-defined or interactively-chosen geographic regions.
+Additionally, **PolarToolkit** provides an easy means for exploring datasets with pre-defined or interactively chosen geographic regions.
 
 <!-- SPHINX-END-proj-desc -->
 
@@ -53,51 +48,35 @@ alt="Latest version on PyPI"/></a>
 🚨 **Ready for daily use but still changing.** 🚨
 </p>
 
-This means that we are still adding a lot of new features and sometimes we make
-changes to the ones we already have while we try to improve the software based
-on users' experience, test new ideas, make better design decisions, etc. Some of
-these changes could be **backwards incompatible**. Keep that in mind before you
-update PolarToolkit to a new major version (i.e. from `v1.0.0` to `v2.0.0`).
+This means that we are still adding a lot of new features and sometimes we make changes to the ones we already have while we try to improve the software based on users' experience, test new ideas, make better design decisions, etc.
+Some of these changes could be **backwards incompatible**.
+Keep that in mind before you update PolarToolkit to a new major version (i.e. from `v1.0.0` to `v2.0.0`) and always check the [Changelog](https://github.com/mdtanker/polartoolkit/blob/main/CHANGELOG.md) for `BREAKING CHANGES` and how to update your code appropriately.
 
-I welcome any feedback, ideas, or contributions! Please contact us on the
-[GitHub discussions page](https://github.com/mdtanker/polartoolkit/discussions)
-or submit an [issue on GitHub](https://github.com/mdtanker/polartoolkit/issues)
-for problems or feature ideas.
+I welcome any feedback, ideas, or contributions!
+Please contact us on the [GitHub discussions page](https://github.com/mdtanker/polartoolkit/discussions) or submit an [issue on GitHub](https://github.com/mdtanker/polartoolkit/issues) for problems or feature ideas.
 
 <!-- SPHINX-START-long-desc -->
 
-The **PolarToolkit** python package provides some basic tools to help in
-conducting polar research. You can use it to download common datasets (i.e.
-BedMachine, Bedmap2, MODIS Mosaics), create maps and plots specific to
-Antarctica, Greenland and the Arctic and visualize data with multiple methods.
+The **PolarToolkit** python package provides some basic tools to help in conducting polar research.
+You can use it to download common datasets (i.e. BedMachine, Bedmap2, MODIS Mosaics), create maps and plots specific to Antarctica, Greenland and the Arctic and visualize data with multiple methods.
 
-Feel free to use, share, modify, and
-[contribute](https://polartoolkit.readthedocs.io/en/latest/contributing.html) to
[…]
