+++ docs/install.md
@@ -2,52 +2,45 @@
 
 ## Online usage (Binder)
 
-See below for the full installation instructions. If instead you'd like to use
-this package online, without needing to install anything, check out our
-[Binder link](https://mybinder.org/v2/gh/mdtanker/polartoolkit-binder/main?urlpath=git-pull%3Frepo%3Dhttps%253A%252F%252Fgithub.com%252Fmdtanker%252Fpolartoolkit%26urlpath%3Dtree%252Fpolartoolkit%252Fdocs%252Fgallery%26branch%3Dmain),
-which gives full access the the package in an online environment.
+See below for the full installation instructions. If instead you'd like to use this package online, without needing to install anything, check out our [Binder link](https://mybinder.org/v2/gh/mdtanker/polartoolkit-binder/main?urlpath=git-pull%3Frepo%3Dhttps%253A%252F%252Fgithub.com%252Fmdtanker%252Fpolartoolkit%26urlpath%3Dtree%252Fpolartoolkit%252Fdocs%252Fgallery%26branch%3Dmain), which gives full access to the latest released version of the package in an online environment.
 
 ## Install Python
 
-Before installing _PolarToolkit_, ensure you have Python downloaded. If you
-don't, I recommend setting up Python with Miniforge. See the install
-instructions [here](https://github.com/conda-forge/miniforge).
+Before installing _PolarToolkit_, ensure you have Python 3.9 or greater downloaded.
+If you don't, I recommend setting up Python with Miniforge.
+See the install instructions [here](https://github.com/conda-forge/miniforge).
 
 ## Install _PolarToolkit_ Locally
 
-There are 3 main ways to install `polartoolkit`. We show them here in order of
-simplest to hardest.
+There are 3 main ways to install `polartoolkit`. We show them here in order of simplest to hardest.
 
 ### Conda / Mamba
-
-The easiest way to install this package and it's dependencies is with conda or
-mamba into a new virtual environment:
-
-```
-mamba create --name polartoolkit --yes --force polartoolkit
-```
-
-Activate the environment:
-
-```
-conda activate polartoolkit
-```
-
-### Pip
-
-Instead, you can use pip to install `polartoolkit`, but first you need to
-install a few dependencies with conda. This is because `PyGMT` `GeoPandas`, and
-`Cartopy` all rely on C packages, which can only be install with conda/mamba and
-not with pip.
 
 ```{note}
 `conda` and `mamba` are interchangeable
 ```
 
+The easiest way to install this package and it's dependencies is with conda or mamba into a new virtual environment:
+
+```
+mamba create --name polartoolkit --yes --force polartoolkit --channel conda-forge
+```
+
+Activate the environment:
+
+```
+mamba activate polartoolkit
+```
+
+### Pip
+
+Instead, you can use pip to install `polartoolkit`, but first you need to install a few dependencies with conda.
+This is because `PyGMT` `GeoPandas`, and `Cartopy` all rely on C packages, which can only be successfully install with conda/mamba and not with pip.
+
 Create a new virtual environment:
 
 ```
-mamba create --name polartoolkit --yes --force pygmt geopandas cartopy
+mamba create --name polartoolkit --yes --force pygmt geopandas cartopy --channel conda-forge
 ```
 
 activate the environment and use `pip` to install `polartoolkit`:
@@ -64,8 +57,7 @@ to install the optional dependencies, use this instead:
 
 ### Development version
 
-You can use pip, with the above created environment, to install the latest
-source from GitHub:
+You can use pip, with the above created environment, to install the latest source from GitHub:
 
     pip install git+https://github.com/mdtanker/polartoolkit.git
 
[…]
