### 2024-06-11 #6502 @mdtanker
Another update: this turned into a slightly bigger task than I had anticipated. Adding support for Greenland and the Arctic required some pretty significant refactoring of the code, but I think it is done now.

Remaining  tasks to do:
- [ ] expand the state of the field in the paper
- [ ] https://github.com/mdtanker/polartoolkit/issues/200
- [ ] release `v0.4.0` by merging https://github.com/mdtanker/polartoolkit/pull/195

I should be able to get these done soon and give line-by-line responses to the suggestions. Thanks again for all of your patience with this!
