### 2024-06-11 04d3ac2
docs: add release instructions to contributing guide
