### 2024-06-11 mdtanker/polartoolkit#200 @mdtanker
Thanks for the suggestions! I had a look, but I'm not sure if this will work here. `polartoolkit` uses `pooch` to download the datasets. For the NASA data, `pooch` requires the earthdata username and password to be entered into the download call. Looking through earthaccess, I couldn't see a way to return the username and password values to pass to pooch. Maybe I missed it, is there  a way to do this?

For now, I've kept it as is, but added better documentation for how to set temporarily or permanent environment variables so users don't need to continuously supply the login credentials.
