### 2024-06-11 085eadf
docs: add instructions for EarthData login
