+++ docs/datasets/greenland/index.md
@@ -17,6 +17,14 @@ geoid
 modis_mog
 ```
 
+## Geophysically-derived data
+
+```{nbgallery}
+---
+---
+mass_change
+```
+
 ## Shapefiles and Misc.
 
 ```{nbgallery}
