### 2024-06-11 a20caa7
chore: update fetch mass change test
