+++ tests/test_fetch.py
@@ -1048,7 +1048,7 @@ def test_magnetics(test_input, expected):
 
 mass_change_test = [
     (
-        "ais_dhdt_floating",
+        "dhdt_floating",
         (
             5001.18699879,
             (-2521652.10412, 2843360.03282, -2229531.47932, 2336552.25058),
@@ -1056,9 +1056,10 @@ mass_change_test = [
             11.7978229523,
             "p",
         ),
+        "south",
     ),
     (
-        "ais_dmdt_grounded",
+        "dmdt_grounded",
         (
             5000.0,
             (-2526157.06916, 2648842.93084, -2124966.01441, 2180033.98559),
@@ -1066,6 +1067,29 @@ mass_change_test = [
             1.0233386755,
             "p",
         ),
+        "south",
+    ),
+    (
+        "dhdt",
+        (
+            5000.0,
+            (-626302.876984, 818697.123016, -3236440.11184, -681440.111839),
+            -58.5592041016,
+            1.15196788311,
+            "p",
+        ),
+        "north",
+    ),
+    (
+        "dmdt_filt",
+        (
+            5000.0,
+            (-626302.876984, 818697.123016, -3236440.11184, -681440.111839),
+            -36.8038787842,
+            0.690221190453,
+            "p",
+        ),
+        "north",
     ),
 ]
 
@@ -1073,9 +1097,9 @@ mass_change_test = [
 @pytest.mark.fetch()
 @pytest.mark.issue()
 @pytest.mark.filterwarnings("ignore::RuntimeWarning")
-@pytest.mark.parametrize(("test_input", "expected"), mass_change_test)
-def test_mass_change(test_input, expected):
-    grid = fetch.mass_change(test_input)
+@pytest.mark.parametrize(("test_input", "expected", "hemisphere"), mass_change_test)
+def test_mass_change(test_input, expected, hemisphere):
+    grid = fetch.mass_change(version=test_input, hemisphere=hemisphere)
     # assert utils.get_grid_info(grid) == pytest.approx(expected, rel=0.1)
     assert not deepdiff.DeepDiff(
         utils.get_grid_info(grid),
