+++ src/polartoolkit/fetch.py
@@ -470,16 +470,23 @@ def basal_melt(variable: str = "w_b") -> typing.Any:
 
 def ice_vel(
     region: tuple[float, float, float, float] | None = None,
-    spacing: float = 5e3,
+    spacing: float | None = None,
     registration: str | None = None,
+    hemisphere: str | None = None,
     **kwargs: typing.Any,
 ) -> xr.DataArray:
     """
-    MEaSUREs Phase-Based Antarctica Ice Velocity Map, version 1 from
-    :footcite:t:`mouginotcontinent2019` and :footcite:t:`mouginotmeasures2019`.
+    MEaSUREs Phase-Based Ice Velocity Maps for Antarctica and Greenland.
+
+    Antarctica: version 1 from :footcite:t:`mouginotcontinent2019` and
+    :footcite:t:`mouginotmeasures2019`.
 
     accessed from https://nsidc.org/data/nsidc-0754/versions/1#anchor-1
     Data part of https://doi.org/10.1029/2019GL083826
+
+    Greenland: version 1 from :footcite:t:`measures2020`
+
+    accessed from https://nsidc.org/data/nsidc-0670/versions/1
 
     Units are in m/yr
 
@@ -491,12 +498,15 @@ def ice_vel(
     region : tuple[float, float, float, float], optional
         region to clip the loaded grid to, in format [xmin, xmax, ymin, ymax], by
         default doesn't clip
-    spacing : float,
-        grid spacing to resample the loaded grid to, by default 5e3, original spacing
-        is 450m
+    spacing : float, optional,
+        grid spacing to resample the loaded grid to, by default is 5km for Antarctica
+        (original data is 450m), and 250m for Greenland
     registration : str, optional
         change registration with either 'p' for pixel or 'g' for gridline registration,
         by default is None.
+    hemisphere : str, optional
+        choose which hemisphere to retrieve data for, "north" or "south", by default
+        None
     kwargs : typing.Any
         additional keyword arguments to pass to resample_grid
 
@@ -509,107 +519,187 @@ def ice_vel(
     ----------
     .. footbibliography::
     """
-    original_spacing = 450
+    hemisphere = utils.default_hemisphere(hemisphere)
 
-    # preprocessing for full, 450m resolution
-    def preprocessing_fullres(fname: str, action: str, _pooch2: typing.Any) -> str:
-        "Load the .nc file, calculate velocity magnitude, save it back"
-        fname1 = Path(fname)
-        # Rename to the file to ***_preprocessed.nc
-        fname_processed = fname1.with_stem(fname1.stem + "_preprocessed_fullres")
-        # Only recalculate if new download or the processed file doesn't exist yet
-        if action in ("download", "update") or not fname_processed.exists():
-            logging.warning(
-                "WARNING; this file is large (~7Gb) and may take some time to download!"
-            )
-            logging.warning(
-                """WARNING; preprocessing this grid in full resolution is very
-                computationally demanding, consider choosing a lower resolution using
-                the parameter `spacing`."
-                """
-            )
-            with xr.open_dataset(fname1) as ds:
-                processed = (ds.VX**2 + ds.VY**2) ** 0.5
-                # Save to disk
-                processed.to_netcdf(fname_processed)
-        return str(fname_processed)
+    if hemisphere == "south":
+        if spacing is None:
+            spacing = 5e3
+        original_spacing = 450
 
-    # preprocessing for filtered 5k resolution
-    def preprocessing_5k(fname: str, action: str, _pooch2: typing.Any) -> str:
-        "Load the .nc file, calculate velocity magnitude, resample to 5k, save it back"
-        fname1 = Path(fname)
-        # Rename to the file to ***_preprocessed_5k.nc
[…]
