### 2024-06-12 dc82f0a
refactor: combine fetch `modis_moa` and `modis_mog` to `modis`
BREAKING CHANGE: make sure to update your code to use the new function `fetch.modis()` and specify MoG vs MoA with parameter hemisphere = "south" or "north"
