+++ src/polartoolkit/fetch.py
@@ -712,66 +712,20 @@ def ice_vel(
     return typing.cast(xr.DataArray, resampled)  # pylint: disable=possibly-used-before-assignment
 
 
-def modis_moa(
-    version: str = "750m",
+def modis(
+    version: str | None = None,
+    hemisphere: str | None = None,
 ) -> str:
     """
-    Load the MODIS Mosaic of Antarctica imagery in either "750m" or "125m" resolutions
+    Load the MODIS Mosaic of Antarctica (MoA) or Greenland (MoG) imagery.
+
+    Antarctica:
     from :footcite:t:`haranmodis2021` and :footcite:t:`scambosmodisbased2007`.
 
     accessed from https://nsidc.org/data/nsidc-0593/versions/2
 
-    Requires an EarthData login, see Tutorials/Download Polar datasets for how to
-    configure this.
-
-    Parameters
-    ----------
-    version : str, optional
-        choose between "750m" or "125m" resolutions, by default "750m"
-
-    Returns
-    -------
-    str
-       filepath for either 750m or 125m MODIS MoA Imagery
-
-    References
-    ----------
-    .. footbibliography::
-    """
-    if version == "125m":
-        path: str = pooch.retrieve(
-            url="https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0593_moa2009_v02/geotiff/moa125_2009_hp1_v02.0.tif.gz",
-            fname="moa125.tif.gz",
-            path=f"{pooch.os_cache('pooch')}/polartoolkit/imagery",
-            downloader=EarthDataDownloader(),
-            processor=pooch.Decompress(method="gzip", name="moa125_2009_hp1_v02.0.tif"),
-            known_hash="101fa22295f94f6eab487d208c051cf81c9af925355b124a04e3d96463af5b72",
-            progressbar=True,
-        )
-    elif version == "750m":
-        path = pooch.retrieve(
-            url="https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0593_moa2009_v02/geotiff/moa750_2009_hp1_v02.0.tif.gz",
-            fname="moa750.tif.gz",
-            path=f"{pooch.os_cache('pooch')}/polartoolkit/imagery",
-            downloader=EarthDataDownloader(),
-            processor=pooch.Decompress(method="gzip", name="moa750_2009_hp1_v02.0.tif"),
-            known_hash="90d1718ea0971795ec102482c47f308ba08ba2b88383facb9fe210877e80282c",
-            progressbar=True,
-        )
-    else:
-        msg = "invalid version string"
-        raise ValueError(msg)
-
-    return path
-
-
-def modis_mog(
-    version: str = "500m",
-) -> str:
-    """
-    Load the 2015 MODIS Mosaic of Greenland imagery in either "500m" or "100m"
-    resolutions from :footcite:t:`haranmeasures2018`.
-
+    Greenland:
+    from :footcite:t:`haranmeasures2018`
     accessed from https://nsidc.org/data/nsidc-0547/versions/2
 
     Requires an EarthData login, see Tutorials/Download Polar datasets for how to
@@ -780,40 +734,82 @@ def modis_mog(
     Parameters
     ----------
     version : str, optional
-        choose between "500m" or "100m" resolutions, by default "500m"
+        for Antarctica, choose between "750m" or "125m" resolutions, by default "750m",
+        for Greenland, choose between "500m" or "100m" resolutions, by default "500m"
+    hemisphere : str, optional
+        choose which hemisphere to retrieve data for, "north" or "south", by default
+        None
 
     Returns
     -------
     str
-       filepath for either 500m or 100m MODIS MoG Imagery
+       filepath for MODIS Imagery
 
[…]
