### 2024-06-12 9876161
build: add deprecation package to dev deps
