### 2024-06-12 7c5461e
chore: add deprecations to old fetch modis functions
