### 2024-06-12 80df7bb
chore: add deprecation to old fetch measures boundaries function
