+++ src/polartoolkit/fetch.py
@@ -33,7 +33,6 @@ from dotenv import load_dotenv
 from pyproj import Transformer
 
 import polartoolkit
-
 from polartoolkit import (  # pylint: disable=import-self
     fetch,  # noqa: PLW0406
     regions,
@@ -818,7 +817,7 @@ def modis(
     deprecated_in="0.4.0",
     removed_in="0.8.0",
     current_version=polartoolkit.__version__,
-    details="Use the new function use modis(hemisphere='south') instead",
+    details="Use the new function modis(hemisphere='south') instead",
 )
 def modis_moa(version: str = "750m") -> str:
     """deprecated function, use modis(hemisphere="south") instead"""
@@ -829,7 +828,7 @@ def modis_moa(version: str = "750m") -> str:
     deprecated_in="0.4.0",
     removed_in="0.8.0",
     current_version=polartoolkit.__version__,
-    details="Use the new function use modis(hemisphere='north') instead",
+    details="Use the new function modis(hemisphere='north') instead",
 )
 def modis_mog(version: str = "500m") -> str:
     """deprecated function, use modis(hemisphere="north") instead"""
@@ -1817,8 +1816,8 @@ def bedmachine(
         change registration with either 'p' for pixel or 'g' for gridline registration,
         by default is None.
     hemisphere : str, optional
-        choose which hemisphere to retrieve data for, currently only available for
-        "south", by default "south"
+        choose which hemisphere to retrieve data for, "north" or "south", by default
+        None
 
     Returns
     -------
@@ -2964,7 +2963,7 @@ def geoid(
                 projection=proj,  # pylint: disable=possibly-used-before-assignment
                 spacing=initial_spacing,
             )
-            # get just antarctica region
+            # get just needed region
             processed = pygmt.grdsample(
                 grid2,
                 region=initial_region,
