### 2024-06-12 4a83f67
build: move deprecation from dev to normal dependencies
