### 2024-06-12 c02fecd
docs: merge arctic and greenland dataset gallery
