### 2024-06-12 d71a64e
docs: merge arctic and greenland dataset gallery
