+++ docs/datasets/arctic/index.md
@@ -1,8 +1,0 @@
-# Arctic
-
-> No datasets added for the Arctic yet!
-
-```{nbgallery}
----
----
-```
