+++ docs/datasets/index.md
@@ -33,11 +33,11 @@ antarctica/basal_melt
 antarctica/groundingline
 antarctica/ice_boundaries
 antarctica/geology
-greenland/modis_mog
-greenland/groundingline
-greenland/bedmachine
-greenland/geoid
-greenland/etopo
-greenland/mass_change
-greenland/ice_velocity
+greenland_arctic/modis_mog
+greenland_arctic/groundingline
+greenland_arctic/bedmachine
+greenland_arctic/geoid
+greenland_arctic/etopo
+greenland_arctic/mass_change
+greenland_arctic/ice_velocity
 ```
