### 2024-06-12 mdtanker/polartoolkit#200 @PennyHow
> For now, I've kept it as is, but added better documentation for how to set temporarily or permanent environment variables so users don't need to continuously supply the login credentials.

I think as long as it is well documented then this is enough to resolve the issue
