### 2024-06-12 aefff40
fix: properly set default region in basemap
