### 2024-06-12 ab78495
feat: add points option to basemap
