### 2024-06-12 ab009d4
fix: remove alter region from get_regions
