### 2024-06-12 f9523a4
docs: major overhaul of docs structure to follow the Divio system
