+++ docs/how_to/index.md
@@ -1,0 +1,15 @@
+# How-To Guides
+```{nbgallery}
+---
+---
+download_NASA_data
+switch_between_hemispheres
+defining_regions
+manipulating_regions
+map_features
+use_with_pygmt
+colormaps
+subplot_layout
+3D_plotting
+profiles
+```
