### 2024-06-12 bbe3975
docs: major overhaul of docs structure to follow the Divio system
