+++ docs/tutorial/index.md
@@ -1,11 +1,16 @@
-# 🚶 Tutorials
+# Tutorials
 
 ```{nbgallery}
 ---
-name: tutorials
 ---
-fetch_walkthrough
-profile_walkthrough
-regions_walkthrough
-utils_walkthrough
+download_data
+specify_download_parameters
+simple_map
+map_features
+subplots
+3D_plots
+plotting_point_data
+cross_sections_and_profiles
+interactive_data
+reprojecting_data
 ```
