### 2024-06-12 0163e20
docs: major overhaul of docs structure to follow the Divio system
