+++ docs/how_to/index.md
@@ -1,4 +1,5 @@
 # How-To Guides
+
 ```{nbgallery}
 ---
 ---
