### 2024-06-12 8f9abc9
fix: fix tests for deprecations
