### 2024-06-13 f87482f
fix: remove hash from mass_change
