+++ src/polartoolkit/fetch.py
@@ -365,8 +365,8 @@ def mass_change(
     path = pooch.retrieve(
         url=url,
         fname=zip_fname,
-        path=f"{pooch.os_cache('pooch')}/polartoolkit/ice_mass",
-        known_hash="8d09ffcce4e84fba8cabbc85ee79fec4de36419f8b242ca1f95adacaa2f229e3",
+        path=f"{pooch.os_cache('pooch')}/polartoolkit/mass_change",
+        known_hash=None,
         progressbar=True,
         processor=pooch.Unzip(
             extract_dir="Smith_2020",
@@ -458,7 +458,7 @@ def basal_melt(variable: str = "w_b") -> typing.Any:
     path = pooch.retrieve(
         url=url,
         fname=fname,
-        path=f"{pooch.os_cache('pooch')}/polartoolkit/ice_mass/Admusilli_2020",
+        path=f"{pooch.os_cache('pooch')}/polartoolkit/mass_change/Admusilli_2020",
         known_hash="c14f7059876e6808e3869853a91a3a17a776c95862627c4a3d674c12e4477d2a",
         progressbar=True,
         processor=preprocessing,
