### 2024-06-13 eaad9c8
fix: fix deprecated test alter region
