### 2024-06-13 c199afa
fix: fix deprecated test measures boundaries
