+++ .github/workflows/ci.yml
@@ -99,3 +99,5 @@ jobs:
 
       - name: Upload coverage report
         uses: codecov/codecov-action@v4.3.0
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
