### 2024-06-13 53e7032
chore: mark some tests as earthdata
