### 2024-06-13 b710923
docs: add logo to README
