### 2024-06-14 2f5771c
chore: add fetch mark to some tests
