### 2024-06-14 090f239
chore: don't run fetch calls during test
This should drastically speed up testing, not overload GHA with downloading datasets, but means testing of fetch calls will need to be performed manual by deleting the local pooch cache and rerun all fetch calls / tests.
