### 2024-06-14 mdtanker/polartoolkit#200 @mdtanker
Ok that sounds good to me, I'll close this as I've added better documentation in #212 in [this how-to guide](https://polartoolkit.readthedocs.io/en/latest/how_to/download_NASA_data.html).
