### 2024-06-15 #6502 @mdtanker
Thank you @JessicaS11, @PennyHow for the really great reviews! Your comments were very helpful and I think this package is much better for it :)

Version `0.4.0` is now available on pip and conda, and it  contains all the fixes and features related to this review. I done a major overhaul on the docs as well, trying to align better with the [Divio system](https://documentation.divio.com/) for documentation, as described a bit in this issue: https://github.com/mdtanker/polartoolkit/issues/209.

See the latest comment from `editorialbot` for the updated version of the paper.

Below I'll address each of your comments line-by-line.

# Jessica's comments
> Hello @mdtanker! Great toolkit. Below please find my review comments, notwithstanding [mdtanker/polartoolkit#193](https://github.com/mdtanker/polartoolkit/issues/193), [mdtanker/polartoolkit#194](https://github.com/mdtanker/polartoolkit/issues/194), [mdtanker/polartoolkit#62](https://github.com/mdtanker/polartoolkit/issues/62), and agreement with review comments from @PennyHow on adding more information to the documentation.

Those 3 issues should all be fixed in `v0.4.0` now!

> ## Documentation
> I really like all the invitations for contributions, including reminders and notes that it's okay if your submission isn't perfect or you don't know how to use all the tools.

Thanks! I found these kinds of comments very encouraging when I was starting to contribute to other packages.

> > Example Usage
>
> Do you have any examples of a scientific workflow (perhaps as part of a tutorial or publication) containing one of the figures generated by this software?

That is a good idea. Currently I don't, but I have some publication in prep now which will use PolarToolkit extensvely. Once these are published, I will add a section to the docs for showcase uses of PolarToolkit.

> > Community guidelines
>
> For updates (or new) notebooks, should contributors run them locally and push with outputs, or clear all outputs?
>

I've updated the contributors guides to have some more details on this: https://polartoolkit.readthedocs.io/en/latest/contributing.html#documentation

> ## Software Paper
> There's a link typo (to the old repo) on line 38.
>

fixed

> > Summary and Statement of Need
>
> The summary is well written, though it suggests the current functionality enables all "Polar" research while in reality it is still limited to Antarctica, with plans to expand to the Arctic/Greenland (yay!). This technicality is important from a user perspective and also because expanding the current package to the north pole will include substantial additions that are not captured in the software release with the current paper.
>

Yes I agree and think I was a little ahead of myself to submit this to JOSS (hence why these revisions have taken me so long!) I have now full implemented Arctic and Greenland support for all the functions including mapping and datasets. You can see the currently available Arctic/Greenland datasets at https://polartoolkit.readthedocs.io/en/latest/datasets/greenland_arctic/index.html.

> It would also be helpful for more details in the paper (and documentation) around what types of datasets are available, and what types of polar research they are enabling. For instance, it appears that most (all?) of the source datasets are gridded.
>

Great suggestion. I've added a section to the README / index describing what PolarToolkit is and isn't intended to do: https://polartoolkit.readthedocs.io/en/latest/index.html#what-polartoolkit-is-for

> > State of the field
>
> Could you please elaborate on how this fits into the broader tool landscape? For instance, as noted in the documentation, the fetch module uses icepack's EarthDataDownloader class.

Yes, I've updated the paper with some more details about how this relates to similar packages. Please let me know if you think my additions are good. I'll include them below for ease:

Written in easy-to-learn Python, and utilizing common geospatial data structures, PolarToolkit is designed to be familiar to use for experienced Python users, while also being approachable for beginner coders. It is built upon several open-source packages, such as [Pooch](https://www.fatiando.org/pooch/latest/) for data downloading [@uiedapooch2020], [PyGMT](https://www.pygmt.org/latest/) for creating figures [@uiedapygmt2021], and [Xarray](https://docs.xarray.dev/en/stable/) and [Verde](https://www.fatiando.org/verde/latest/) for geospatial data processing [@hoyerxarray2017; @uiedaverde2018]. PolarToolkit is designed for generic work with polar data while there are specific packages available for downloading or working with specific data types, such as [ITS_LIVE](https://github.com/nasa-jpl/itslive-py) for glacier velocity data [@itslive2024], [icepyx](https://github.com/icesat2py/icepyx) for ICESat-2 data [@scheickicepyx2023], and [earthspy](https://github.com/AdrienWehrle/earthspy) for satellite data [@earthspy2024]. Other similar software exists but differ from PolarToolkit in being either focused on specific types of data [[icepack](https://github.com/icepack/icepack); @shaperoicepack2023] or using different programming languages which require users to have paid licenses [Antarctic Mapping Tools, @greeneantarctic2017]. Comprehensive documentation, API reference, tutorials, and how-to guides are available at https://polartoolkit.readthedocs.io/en/, and development occurs in the [GitHub repository](https://github.com/mdtanker/polartoolkit).

# Penny's comments

> Thanks for a nice submission, @mdtanker! Generally I feel that this is a great package with good entry-level functionality for beginners in programming, and also handy tools for seasoned cryospheric coders to quickly generate publication-ready figures. The documentation and package structuring reflects these two target user types, with thorough documentation and tutorials.
>

Thanks alot!

> I have two major comments that need to be addressed before accepting, along with some minor comments:
>
> ## The functionality is limited to Antarctica
> The term "polar" places an expectation on the package being usable with Antarctic and Arctic datasets. Currently functionality only accomodates datasets for Antarctica. You say that ideally you would like polartoolkit to develop further for use with data from Greenland and the Arctic. On the readthedocs home page, it is stated that this is happening "soon". When is "soon" and would this be within the timeframe of this review process?
>

Please see above comment to Jessica's review

> Whilst I do not need to see a full implementation of these Greenland/Arctic developments, I think there should be certain changes in the naming conventions and package structuring so that it is easy to foresee how Greenland/Arctic datasets and analysis could be incorporated. For example, I see that the names of the currently available datasets are generically named and not place-specific:
>
>
> These should be changed accordingly so that Greenland/Arctic dataset names could be easily incorporated, i.e. basal_melt >> antarctic_basal_melt.
>
> If you think it is possible, I would also like to request one Greenland/Arctic example be developed and placed in the tutorials/gallery to demonstrate the viability of this future development.
>

To address this I've done a few things:
* all functions which required knowing which hemishere (north for Arctic/Greenland or south for Antarctic) know have a `hemisphere` variable which can be set equal to either `north` or `south`.  This can be done manually, via a temporary python environment variable (os.environ[POLARTOOLKIT_HEMISPHERE]="north") or via a permanent system environment variable (i.e. within .bashrc). There are detailed instructions for this [here](https://polartoolkit.readthedocs.io/en/latest/how_to/switch_between_hemispheres.html).
     * this allows researchers only focused on 1 region to set this parameter and forget (via the system environment variable), or researchers who flip between to have each notebook set the hemisphere at the top (via python environment variables).
* for projections, north hemisphere relates to EPSG 3413 and south relates to EPSG 3031.
* datasets where were vaguely names have been updated (i.e. measures_boudaries -> antarctic_boundaries)
* datasets which have both north and south hemispheres avaible (e.g. bedmachine, ice_velocity, ETOPO) now take the `hemisphere` argument to clarify which to return.

Let me know if you think this is a decent way of implementing this.

> ## The installation needs to be more straightforward for beginner users
> The extensive list of package dependencies mean that it takes a long time for mamba/conda to solve the environment with installation. Even the Binder online build did not work for me as it cannot solve the environment in a timely manner.
>
> As this package is partially aimed at beginners in programming, the set-up needs to be simple otherwise it is not accessible for these users. Whilst the installation troubleshooting documentation is good, I think that the package should be modified to make the installation easier. You can solve this problem however you think is best, whether it be changing the package distribution itself, or adding more extensive installation guidance.

I've made the [install instructions ](https://polartoolkit.readthedocs.io/en/latest/install.html) clearer I think, and included an `environment.yml` file in the repo as a back option to install.

>
> Suggestion: I would look at the number of dependencies and refine them by trying to find common functionality between them. And I would also look at alternatives to geopandas, cartopy and pygmt that do not have C bindings (as this is may resolve the issues with the online Binder build).
>

This is a tricky one which I've been trying to solve since early on in PolarToolkit. For the three C packages, PyGMT is obviously needed, I can't seem to find an alternative to geopandas for reading in shapefiles (I've tried pyogrio to no success (https://github.com/mdtanker/polartoolkit/issues/18), and cartopy is currently an optional dependency. Of the other dependencies, unfortunately I think they are all necessary. I have updated the binder build to be based on a seperate repository (so it is less-frequently rebuilt) and simplified it's install, hopefully this helps out. (I need to wait a few more hours for conda to release `v0.4.0` before I can update the binder env by the way.

> ## Minor comments
> **Functionality**
>
> * [x]  [Indentation error in subplots tutorial mdtanker/polartoolkit#198](https://github.com/mdtanker/polartoolkit/issues/198)

Fixed

> * [x]  [Plot viewing procedures differ in the Download Antarctica datasets tutorial mdtanker/polartoolkit#199](https://github.com/mdtanker/polartoolkit/issues/199)

Fixed

> * [x]  [EarthDataDownloader object to retain user login credentials mdtanker/polartoolkit#200](https://github.com/mdtanker/polartoolkit/issues/200)

Fixed (with documentation)

>
> **Installation**
>
> * [ ]  The repo is quite big for cloning. Is there anything you can remove to slim down the repo contents and size?

I've moved the binder env out. I also removed the `data/` folder, which had 2 example shapefiles to use in making profiles.  These are now stored in a FigShare repo and downloaded with Pooch. I don't think there's anything else I'm able to delete.

> * [ ]  For conda installation, please define that users should make sure they should add the conda-forge channel

Added

> * [ ]  Can you provide some install tests for users in the documentation and repo readme to ensure that their installation works. It could be a simple version call, or some form of unit tests.

I added some instructions to print the version after install as a simple check for a correct install. I've left the tests folder out of the build to try and reduce size, but this unfortunately means users can run the test, but I don't think that is very common for users to do anyways.

> * [ ]  Please can you add the environment yml file to the repo which you previously provided me. I think it is a good option to have if a user is troubleshooting their install (like me!)
>

Yes this is now included, and is mentioned in the install instructions as a backup method.

> **Documentation**
>
> * [x]  Most of the issues with documentation have already been addressed in [this PR](https://github.com/mdtanker/polartoolkit/pull/197)
>

> **Software paper**
>
> * [ ]  Regarding the state of the field: I think there are other resources/commonly-used packages out there for arctic data handling - check out the [cryosphere awesome list](https://github.com/awesome-cryosphere/cryosphere-links). I would include earthspy at least. I see you also reference [Fatiando](https://www.fatiando.org/) in the readthedocs documentation, which should probably be added here also.
>

Thanks for this suggestion, please see my response above  to Jessica's similar comment.

> **Other**
>
> * [ ]  The logo for polartoolkits looks fairly similar to other software logos I have come across, such as [this](https://github.com/RJArthern/WAVI.jl). It's not really anything that needs specifically changing (maybe it is hard to avoid!), but I just thought I would make you aware that this type of logo is quite common for software related to cryospheric sciences.

Intersting, yeah after looking there do seem to be some similar ones ... I think I'll stick with this as it took some time to make and I think it turned out alright, but maybe I should have done something more unique!
