### 2024-06-21 mdtanker/polartoolkit#200 @JessicaS11
> Looking through earthaccess, I couldn't see a way to return the username and password values to pass to pooch. Maybe I missed it, is there a way to do this?

It's not part of the user api (and probably not documented), but with an earthaccess Auth object (created with `myauth = earthaccess.login()`) you should be able to get them with [`myauth.auth_environ()`](https://github.com/nsidc/earthaccess/blob/9563af0cc8ad39ec676c286ec27c7adbb146059a/earthaccess/api.py#L364) (or possibly `myauth.auth.auth_environ()`).

As @PennyHow noted - not required for JOSS, but wanted to follow up and answer your question!
