### 2024-06-22 mdtanker/polartoolkit#200 @mdtanker
Ahh great! Thanks for finding that, this is working to access the credentials: `earthaccess.auth_environ()`.

I'll add this into PolarToolkit to replace the custom class I had.
