### 2024-06-24 #6502 @JessicaS11
Great updates, @mdtanker! It's awesome that you were able to implement the northern hemisphere, and the updated documentation around available datasets is great. Is there an easy way to include the data source within the fetched object itself (e.g. as an attribute)? I found myself wanting to view this info programmatically next to the data without backtracing up to my initial `fetch` to see which dataset I had in memory.

My primary outstanding concern:
- after the most recent round of updates, many of the [make] commands in the "How to Contribute" section (including tests) are no longer runnable by the user without further setup.

Minor comments and notes:
- what "specific types" of data does icepack focus on (line 37)?
- the paper has a few minor typos (e.g. python not capitalized, singular/plural mismatch)

> I've moved the binder env out. I also removed the data/ folder, which had 2 example shapefiles to use in making profiles. These are now stored in a FigShare repo and downloaded with Pooch. I don't think there's anything else I'm able to delete.

I suspect that removing the shapefiles did not substantially decrease the size of the repo (and unless the history was rewritten, would not change the repo size at all anyway). My guess is that the bulk of the repo size is due to the storage of the notebook outputs.
