### 2024-07-02 #6502 @mdtanker
> Great updates, @mdtanker! It's awesome that you were able to implement the northern hemisphere, and the updated documentation around available datasets is great. Is there an easy way to include the data source within the fetched object itself (e.g. as an attribute)? I found myself wanting to view this info programmatically next to the data without backtracing up to my initial `fetch` to see which dataset I had in memory.

That is a good idea. I've played around with it and if I switched all the fetch functions to classes, each could have attributes like DOI's, descriptions, and source URL's. I think this would be a good feature so I'm happy to implement this, but it will definently take some time to switch all the functions to classes. You can see the initial progress here: https://github.com/mdtanker/polartoolkit/actions/runs/9764818494/job/26954042687?pr=217

Is this something you would like to see as part of this review or just a feature to be implemented at some point?

> My primary outstanding concern:
> - after the most recent round of updates, many of the [make] commands in the "How to Contribute" section (including tests) are no longer runnable by the user without further setup.

Yes that is a good point. I have added some instructions on setting up `make` for contributors.

>
> Minor comments and notes:
> - what "specific types" of data does icepack focus on (line 37)?
> - the paper has a few minor typos (e.g. python not capitalized, singular/plural mismatch)

Thanks, I've update the paper to fix the typos I found, and change the icepack statement to be more along the lines of specific uses (ice flow modelling) than about data types.

> > I've moved the binder env out. I also removed the data/ folder, which had 2 example shapefiles to use in making profiles. These are now stored in a FigShare repo and downloaded with Pooch. I don't think there's anything else I'm able to delete.
>
> I suspect that removing the shapefiles did not substantially decrease the size of the repo (and unless the history was rewritten, would not change the repo size at all anyway). My guess is that the bulk of the repo size is due to the storage of the notebook outputs.

I think you're right, but due to the issues with the C dependencies, I can't run many of the notebooks within Read the Docs, and so I need to run them locally and commit the outputs. I have slightly reduced the size of the `docs/` folder by adding the `PyGMT` parameter `dpi=200` to every `fig.show()` call in each notebook. I think this has reduced the size a bit, but the figures are still an acceptable quality. If you have any ideas for further reducing the size please let me know!

All of the above changes are in this PR: https://github.com/mdtanker/polartoolkit/pull/216
