### 2024-07-12 #6502 @PennyHow
Hi @mdtanker, sorry it's taken me a while to provide some feedback. After looking through your amendments, I can see that you have really spent a lot of time expanding the scope of PolarToolkit and refining it based on our comments. Very nice job!

Just two comments from me:

- Taking the binder build out of the repo has definitely helped with the repo cloning. And I understand regarding the dependencies. I think slimming down the repo as much as possible, and making additional installation documentation is a viable solution

- The expansion with Greenland datasets is really great to see, and definitely now fulfills the "polar" name. I think your approach to defining a `hemisphere` variable works well, and the instructions guide the user very clearly. One question I have: Is there a reason why exclusively US datasets are used (e.g. MEaSUREs)? I don't think you need to change anything, but maybe a thought for future versions would be to expand to incorporate other data sources... European datasets, for example, are pretty good too!
