### 2024-07-17 #6502 @hugoledoux
@mdtanker "Is this something you would like to see as part of this review or just a feature to be implemented at some point?" from https://github.com/openjournals/joss-reviews/issues/6502#issuecomment-2204107690
==> I guess if it's a lot of work then you could leave it for the future, we can accept the submission without this.

@JessicaS11 when you have time, could you confirm that the latest changes (https://github.com/mdtanker/polartoolkit/pull/216) would fulfil your issues (and can you check the checklist accordingly?).
