### 2024-08-08 #6502 @JessicaS11
Wow! @mdtanker, it's been fun to see the growth of PolarToolkit throughout this review. I'm excited to see this work published in JOSS. Thanks for your patience in between rounds of review.

> Is this something you would like to see as part of this review or just a feature to be implemented at some point?

This is not a specific requirement for this review - I guess it's more of a "user feature request".

> Yes that is a good point. I have added some instructions on setting up make for contributors.

Thanks, @mdtanker, for adding these great instructions (FYI there is some funny rendering on the last line of them). I also discovered today that I had somehow missed running `make install`, which is why I was still missing all the dependencies for the other make commands. I re-followed the instructions and everything worked fine.

> Thanks, I've update the paper to fix the typos I found, and change the icepack statement to be more along the lines of specific uses (ice flow modelling) than about data types.

Awesome! Thanks for the typo pass and clarification.

@hugoledoux I've completed the checklist. This is good to go.
