### 2024-08-12 #6502 @mdtanker
Thankyou @JessicaS11! I appreciate all the comments and recommendations! I'll keep working on transitioning those fetch functions to methods, and maybe it will be the goal for `v1.0`.  I'll fix those rendering issues.

I'm away tomorrow for a few weeks but will have sporadic internet access. I'm happy with the state of the paper. Please let me know what else there is to be done before this is published!
