### 2024-08-16 #6502 @hugoledoux
@mdtanker not much needs to be done, every is happy at this point. Just please accept this PR fixing a few minor things, and then:
