### 2024-08-17 #6502 @mdtanker
Alright the released version is v0.5.1 and is on Zenodo at this link: https://doi.org/10.5281/zenodo.13218540

I added the title to the Zenodo and update the author info.
