### 2024-08-20 #6502 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
