### 2024-08-22 #6502 @kthyng
@mdtanker I am delighted by your histogram colorbar! I would love to see that available in a more general toolbox, something like matplotlib! Might you consider that?
