### 2024-08-22 #6502 @mdtanker
Yes I'd love to see that implemented in matplotlib! I'll have a look sometime at their code and how to implement it. I find it so useful to show the distribution!
