### 2024-08-22 #6502 @kthyng
@mdtanker If you're interested in joining JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/!
