### 2024-08-24 #6502 @PennyHow
> Wohoo! Thanks for all the time and energy everyone! I enjoyed this review process and think PolarToolkit is much better due to it! I've signed up to be a reviewer 🙂

@mdtanker, I think you are a great addition to the team of reviewers behind JOSS.Welcome!
